RemoteSensingToolbox

RemoteSensingToolbox is a pure Julia package intended to provide a collection of commonly used tools for working with remotely sensed imagery.

Sensors

Sensors are julia structs that wrap a typical Rasters.RasterStack object to provide compatability with many RemoteSensingToolbox algorithms and methods.

The following methods are supported by all AbstractSensor types:

Base.getindex	return the layer correspinding to the given band name.

Base.length	return the number of layers in the enclosed Rasters.RasterStack.

Base.map	apply a function to each layer in the enclosed Rasters.RasterStack.

Base.write	write layers to file.

Rasters.resample	resample data to a different size and projection, or snap to another object.

Rasters.crop	shrink objects to specific dimension sizes or the extent of another object.

Rasters.extend	extend objects to specific dimension sizes or the extent of another object.

Rasters.trim	trims areas of missing values for arrays and across stack layers.

Rasters.mask	mask an object by a polygon or Raster along X/Y, or other dimensions.

Rasters.replace_missing	replace all missing values in an object and update missingval.

Additionally, asraster can be used to apply a function to the enclosed Rasters.RasterStack.

RemoteSensingToolbox.Sensors.AbstractSensor — Type

The supertype of all sensor types.

Subtypes should wrap a RasterStack under the field 'stack' and implement the following interface:

blue(X::Sensor)

green(X::Sensor)

red(X::Sensor) 

nir(X::Sensor) 

swir1(X::Sensor)

swir2(X::Sensor)

dn2rs(::Type{<:AbstractSensor})

BandSet(::Type{Landsat8})

Example Implementation

struct Landsat8 <: AbstractSensor
    stack::RasterStack
end

function BandSet(::Type{Landsat8})
    bands = [:B1, :B2, :B3, :B4, :B5, :B6, :B7]
    wavelengths = [440, 480, 560, 655, 865, 1610, 2200]
    return BandSet(bands, wavelengths)
end
    
blue(X::Landsat8) = X[:B2]

green(X::Landsat8) = X[:B3]

red(X::Landsat8) = X[:B4]

nir(X::Landsat8) = X[:B5]

swir1(X::Landsat8) = X[:B6]

swir2(X::Landsat8) = X[:B7]

dn2rs(::Type{Landsat8}) = (scale=0.0000275, offset=-0.2)

RemoteSensingToolbox.Sensors.BandSet — Type

A struct for storing the band names and associated wavelengths of a particular sensor.

It is expected that instances of AbstractSensor implement a BandSet constructor.

The central wavelength for a given band can be recovered by calling the BandSet.

Example

julia> bandset = BandSet(Sentinel2A);
julia> bandset(:B8A)
842.0

RemoteSensingToolbox.Sensors.DESIS — Type

stack::Rasters.RasterStack

Implements the AbstractSensor interface for DESIS.

RemoteSensingToolbox.Sensors.Landsat7 — Type

stack::Rasters.RasterStack

Implements the AbstractSensor interface for Landsat 7.

RemoteSensingToolbox.Sensors.Landsat8 — Type

stack::Rasters.RasterStack

Implements the AbstractSensor interface for Landsat 8.

RemoteSensingToolbox.Sensors.Sentinel2A — Type

stack::Rasters.RasterStack

Implements the AbstractSensor interface for Sentinel-2A.

RemoteSensingToolbox.Sensors.asraster — Method

asraster(f, X::AbstractSensor)

Operate on the AbstractSensor as if it was a regular Rasters.RasterStack.

Example

landsat = Landsat8("LC08_L2SP_043024_20200802_20200914_02_T1/")
asraster(landsat) do stack
    map(x -> x .* 0.0001f0, stack)
end

RemoteSensingToolbox.Sensors.blue — Method

blue(X::AbstractSensor)

Return the blue band for the given sensor.

RemoteSensingToolbox.Sensors.dn2rs — Method

dn2rs(::Type{<:AbstractSensor})

Return the scale and offset required to convert DN to reflectance for the given sensor type.

Example

julia> dn2rs(Landsat8)
(scale = 2.75e-5, offset = -0.2)

RemoteSensingToolbox.Sensors.green — Method

green(X::AbstractSensor)

Return the green band for the given sensor.

RemoteSensingToolbox.Sensors.nir — Method

nir(X::AbstractSensor)

Return the nir band for the given sensor.

RemoteSensingToolbox.Sensors.red — Method

red(X::AbstractSensor)

Return the red band for the given sensor.

RemoteSensingToolbox.Sensors.swir1 — Method

swir1(X::AbstractSensor)

Return the swir1 band for the given sensor.

RemoteSensingToolbox.Sensors.swir2 — Method

swir2(X::AbstractSensor)

Return the swir2 band for the given sensor.

Visualization

RemoteSensingToolbox.Agriculture — Type

Agriculture band composite.

RemoteSensingToolbox.ColorInfrared — Type

Color infrared band composite.

RemoteSensingToolbox.Geology — Type

Geology band composite.

RemoteSensingToolbox.SWIR — Type

SWIR band composite.

RemoteSensingToolbox.TrueColor — Type

True color band composite.

RemoteSensingToolbox.visualize — Method

visualize(r::AbstractRaster, g::AbstractRaster, b::AbstractRaster; lower=0.02, upper=0.98)
visualize(g::AbstractRaster; lower=0.02, upper=0.98)
visualize(img::AbstractSensor, ::Type{TrueColor}; lower=0.02, upper=0.98)
visualize(img::AbstractSensor, ::Type{ColorInfrared}; lower=0.02, upper=0.98)
visualize(img::AbstractSensor, ::Type{SWIR}; lower=0.02, upper=0.98)
visualize(img::AbstractSensor, ::Type{Agriculture}; lower=0.02, upper=0.98)
visualize(img::AbstractSensor, ::Type{Geology}; lower=0.02, upper=0.98)

Visualize a remotely sensed image by applying a histogram stretch. Returns either an RGB or grayscale image compatible with the Images.jl ecosystem.

A number of band combinations are supported for types implementing the AbstractSensor interface.

Example 1

landsat = Landsat8("LC08_L2SP_043024_20200802_20200914_02_T1/")
img = visualize(red(landsat), green(landsat), blue(landsat))
save("truecolor.png", img)

Example 2

landsat = Landsat8("LC08_L2SP_043024_20200802_20200914_02_T1/")
img = visualize(landsat, TrueColor)
save("truecolor.png", img)

Preprocessing

RemoteSensingToolbox.Preprocessing.create_tiles — Method

create_tiles(raster, tile::Tuple{Int,Int}; stride=tile)

Slice the given raster into tiles with size tile.

Parameters

raster: The raster to be cut into tiles.

tile: The size of the generated tiles in terms of width x height.

stride: The distance between the top-left corner of each tile. Is equal to tile by default, which produces non-overlapping tiles.

RemoteSensingToolbox.Preprocessing.dn_to_reflectance — Method

dn_to_reflectance(rs::AbstractSensor; precision=Float32)
dn_to_reflectance(rs::AbstractRasterStack, scale, offset; precision=Float32)

Transform the raster from Digital Numbers (DN) to reflectance.

Parameters

rs: The RasterStack or AbstractSensor to be converted to reflectance.

scale: The scaling factor used to convert DN to reflectance. Inferred for AbstractSensor types.

offset: The offset used to convert DN to reflectance. Inferred for AbstractSensor types.

precision: The floating point representation of reflectance values. Uses 32 bits of precision by default.

RemoteSensingToolbox.Preprocessing.landsat_qa — Method

landsat_qa(qa_src::String)

Read and decode a landsat quality assurance (QA) raster. Decodes each bit into its own RasterStack layer.

Example

julia> qa = landsat_qa("LC08_L2SP_043024_20200802_20200914_02_T1_QA_PIXEL.TIF")
RasterStack with dimensions: 
  X Projected{Float64} LinRange{Float64}(493785.0, 728385.0, 7821) ForwardOrdered Regular Points crs: WellKnownText,
  Y Projected{Float64} LinRange{Float64}(5.84638e6, 5.60878e6, 7921) ReverseOrdered Regular Points crs: WellKnownText
and 7 layers:
  :dilated_cloud UInt8 dims: X, Y (7821×7921)
  :cirrus        UInt8 dims: X, Y (7821×7921)
  :cloud         UInt8 dims: X, Y (7821×7921)
  :cloud_shadow  UInt8 dims: X, Y (7821×7921)
  :snow          UInt8 dims: X, Y (7821×7921)
  :clear         UInt8 dims: X, Y (7821×7921)
  :water         UInt8 dims: X, Y (7821×7921)

RemoteSensingToolbox.Preprocessing.mask_pixels — Method

mask_pixels(raster, mask; invert_mask=false)

Drop pixels from a raster according to a given mask. The mask and raster must have the same extent and size.

Parameters

raster: The raster to be masked.

mask: A mask defining which pixels we want to drop. By default, we drop pixels corresponding to mask values of 1.

invert_mask: Treat mask values of 1 as 0 and vice-versa.

RemoteSensingToolbox.Preprocessing.tocube — Method

tocube(rs::RasterStack; layers=names(rs))
tocube(rs::AbstractSensor; layers=names(rs))

Transform the multi-layer RasterStack to a multi-band raster.

Parameters

X: The RasterStack or AbstractSensor to be transformed into a multi-band raster.

layers: The layers to include in the new raster.

Example

julia> landsat = Landsat8("LC08_L2SP_043024_20200802_20200914_02_T1");
julia> tocube(landsat)
7821×7921×7 Raster{Float32,3} B1 with dimensions: 
  X Projected{Float64} LinRange{Float64}(493785.0, 728385.0, 7821) ForwardOrdered Regular Points crs: WellKnownText,
  Y Projected{Float64} LinRange{Float64}(5.84638e6, 5.60878e6, 7921) ReverseOrdered Regular Points crs: WellKnownText,
  Band Categorical{Int64} 1:7 ForwardOrdered
extent: Extent(X = (493785.0, 728385.0), Y = (5.608785e6, 5.846385e6), Band = (1, 7))
missingval: 0.0f0
crs: PROJCS["WGS 84 / UTM zone 11N",GEOGCS["WGS 84",DATUM["WGS_1984",SPHEROID["WGS 84",6378137,298.257223563,AUTHORITY["EPSG","7030"]],AUTHORITY["EPSG","6326"]],PRIMEM["Greenwich",0,AUTHORITY["EPSG","8901"]],UNIT["degree",0.0174532925199433,AUTHORITY["EPSG","9122"]],AUTHORITY["EPSG","4326"]],PROJECTION["Transverse_Mercator"],PARAMETER["latitude_of_origin",0],PARAMETER["central_meridian",-117],PARAMETER["scale_factor",0.9996],PARAMETER["false_easting",500000],PARAMETER["false_northing",0],UNIT["metre",1,AUTHORITY["EPSG","9001"]],AXIS["Easting",EAST],AXIS["Northing",NORTH],AUTHORITY["EPSG","32611"]]
parent:
[:, :, 1]
           5.84638e6  5.84636e6  5.84632e6  5.8463e6  5.84626e6  …  5.60894e6  5.6089e6  5.60888e6  5.60884e6  5.60882e6  5.60878e6
 493785.0  0.0        0.0        0.0        0.0       0.0           0.0        0.0       0.0        0.0        0.0        0.0
      ⋮                                               ⋮          ⋱                                             ⋮          
 728355.0  0.0        0.0        0.0        0.0       0.0           0.0        0.0       0.0        0.0        0.0        0.0
 728385.0  0.0        0.0        0.0        0.0       0.0           0.0        0.0       0.0        0.0        0.0        0.0
[and 6 more slices...]

Land Cover Indices

RemoteSensingToolbox.Algorithms.mndwi — Method

mndwi(green::AbstractRaster, swir::AbstractRaster)
mndwi(sensor::AbstractSensor)

Compute the Modified Normalised Difference Water Index (Xu 2006).

MNDWI = (green - swir) / (green + swir)

RemoteSensingToolbox.Algorithms.nbri — Method

nbri(nir::AbstractRaster, swir2::AbstractRaster)
nbri(sensor::AbstractSensor)

Compute the Normalized Burn Ratio Index.

NBRI is used to emphasize burned areas.

NBRI = (nir - swir2) / (nir + swir2)

RemoteSensingToolbox.Algorithms.ndbi — Method

ndbi(swir1::AbstractRaster, nir::AbstractRaster)
ndbi(sensor::AbstractSensor)

Compute the The Normalized Difference Built-up Index

NDBI is used to emphasize urban and built-up areas.

NDBI = (swir1 - nir) / (swir1 + nir)

RemoteSensingToolbox.Algorithms.ndmi — Method

ndmi(nir::AbstractRaster, swir1::AbstractRaster)
ndmi(sensor::AbstractSensor)

Compute the Normalized Difference Moisture Index.

NDMI is sensitive to the moisture levels in vegetation. It is used to monitor droughts and fuel levels in fire-prone areas.

NDMI = (nir - swir1) / (nir + swir1)

RemoteSensingToolbox.Algorithms.ndvi — Method

ndvi(nir::AbstractRaster, red::AbstractRaster)
ndvi(sensor::AbstractSensor)

Compute the Normalized Difference Vegetation Index.

NDVI = (nir - red) / (nir + red)

RemoteSensingToolbox.Algorithms.ndwi — Method

ndwi(green::AbstractRaster, nir::AbstractRaster)
ndwi(sensor::AbstractSensor)

Compute the Normalized Difference Water Index (McFeeters 1996).

NDWI = (green - nir) / (green + nir)

RemoteSensingToolbox.Algorithms.savi — Method

savi(nir::AbstractRaster, red::AbstractRaster; L=0.33)
savi(sensor::AbstractSensor; L=0.33)

Compute the Soil Adjusted Vegetation Index (Huete 1988).

SAVI is a vegetation index which attempts to minimize soil brightness influences by introducing a soil-brightness correction factor (L).

L represents the amount of green vegetation cover, which is set to 0.33 by default.

SAVI = ((nir - red) / (nir + red + L)) * (1 + L)

Spectral Analysis

RemoteSensingToolbox.Spectral.labelled_signatures — Method

labelled_signatures(rs::AbstractSensor, shp::DataFrame, label::Symbol)
labelled_signatures(rs::RasterStack, shp::DataFrame, label::Symbol)

Extract signatures from the given RasterStack or AbstractSensor within regions specified by a given shapefile.

Parameters

rs: The RasterStack or AbstractSensor from which to extract spectral signatures.

shp: A shapefile stored as a DataFrame with a :geometry column storing a GeoInterface.jl compatible geometry and a label column indicating the land cover type.

label: The column in shp in which the land cover class is stored.

Returns

A DataFrame consisting of rows for each extracted signature and columns storing the respective bands and land cover type.

Example

julia> landsat = Landsat8("data/LC08_L2SP_043024_20200802_20200914_02_T1/") |> dn_to_reflectance;

julia> shp = Shapefile.Table("data/landcover/landcover.shp") |> DataFrame;

julia> labelled_signatures(landsat, shp, :cover)
2860×8 DataFrame
  Row │ B1         B2         B3         B4         B5        B6        B7        label      
      │ Float32    Float32    Float32    Float32    Float32   Float32   Float32   String     
──────┼──────────────────────────────────────────────────────────────────────────────────────
    1 │ 0.0514875  0.073845   0.119742   0.128817   0.2466    0.256005  0.221905  Built Up
    2 │ 0.0759625  0.10228    0.152963   0.170865   0.27575   0.300115  0.26387   Built Up
    3 │ 0.113747   0.134263   0.212747   0.237497   0.32756   0.32041   0.28851   Built Up
    4 │ 0.09359    0.117955   0.18522    0.19743    0.302453  0.3148    0.277125  Built Up
    5 │ 0.0793725  0.104507   0.133658   0.1377     0.311555  0.2873    0.23824   Built Up
    6 │ 0.0641375  0.086      0.11724    0.115315   0.353272  0.291755  0.221878  Built Up
    7 │ 0.240688   0.271735   0.254245   0.299648   0.376895  0.500067  0.460797  Built Up
    8 │ 0.06554    0.0865225  0.14265    0.16245    0.272478  0.283697  0.238515  Built Up
    9 │ 0.0979625  0.131732   0.190747   0.230402   0.312793  0.369112  0.327148  Built Up
  ⋮   │     ⋮          ⋮          ⋮          ⋮         ⋮         ⋮         ⋮          ⋮
 2853 │ 0.0355925  0.0422475  0.0688675  0.0711225  0.24385   0.287245  0.180792  Bare Earth
 2854 │ 0.0494525  0.05674    0.0867975  0.09612    0.277703  0.311665  0.192728  Bare Earth
 2855 │ 0.0473075  0.053      0.0815725  0.0860825  0.264998  0.291672  0.183075  Bare Earth
 2856 │ 0.04673    0.05267    0.08039    0.083195   0.25716   0.285787  0.184175  Bare Earth
 2857 │ 0.0465925  0.05289    0.0809675  0.0854225  0.268627  0.299345  0.189565  Bare Earth
 2858 │ 0.0471975  0.052835   0.08094    0.0856975  0.266097  0.302783  0.189097  Bare Earth
 2859 │ 0.0453275  0.050745   0.075055   0.0795925  0.247095  0.293323  0.185962  Bare Earth
 2860 │ 0.036335   0.0434025  0.06884    0.070985   0.2422    0.282845  0.179995  Bare Earth
                                                                            2843 rows omitted

RemoteSensingToolbox.Spectral.plot_signatures! — Method

plot_signatures!(ax::Axis, rs::AbstractSensor, shp::DataFrame, label::Symbol; colors=wong_colors())
plot_signatures!(ax::Axis, rs::RasterStack, shp::DataFrame, bandset::BandSet, label::Symbol; colors=wong_colors())

Plot spectral signatures for each land cover type specified in a given shapefile by mutating a Makie.Axis object.

Parameters

ax: The Makie.Axis into which we want to draw our plot.

rs: The RasterStack or AbstractSensor from which to extract spectral signatures.

shp: A shapefile stored as a DataFrame with a :geometry column storing a GeoInterface.jl compatible geometry and a label column indicating the land cover type.

bandset: The BandSet for the provided sensor specifying the available bands and associated wavelengths in nm. Inferred for AbstractSensor.

label: The column in shp in which the land cover class is stored.

colors: The color scheme used by the plot.

Example

# Read Landsat And Convert DNs To Reflectance
landsat = Landsat8("data/LC08_L2SP_043024_20200802_20200914_02_T1/") |> dn_to_reflectance

# Load Shapefile
shp = Shapefile.Table("data/landcover/landcover.shp") |> DataFrame

# Create Axes
fig = Figure();
ax1 = Axis(fig[1,1]);
ax2 = Axis(fig[2,1]);

# Plot Signatures
plot_signatures!(ax1, landsat, shp, :C_name; colors=cgrad(:tab10));
plot_signatures!(ax2, landsat, shp, :MC_name; colors=cgrad(:tab10));

# Add Legend
Legend(fig[1,2], ax1);
Legend(fig[2,2], ax2);

# Save Figure
save("landsat_signatures.png", fig)

RemoteSensingToolbox.Spectral.plot_signatures — Method

plot_signatures(rs::AbstractSensor, shp::DataFrame, label::Symbol; colors=wong_colors())
plot_signatures(rs::RasterStack, shp::DataFrame, bandset::BandSet, label::Symbol; colors=wong_colors())

Plot spectral signatures for each land cover type specified in a given shapefile.

Parameters

rs: The RasterStack or AbstractSensor from which to extract spectral signatures.

shp: A shapefile stored as a DataFrame with a :geometry column storing a GeoInterface.jl compatible geometry and a label column indicating the land cover type.

bandset: The BandSet for the provided sensor specifying the available bands and associated wavelengths in nm. Inferred for AbstractSensor.

label: The column in shp in which the land cover class is stored.

colors: The color scheme used by the plot.

Example

# Read Landsat And Convert DNs To Reflectance
landsat = Landsat8("data/LC08_L2SP_043024_20200802_20200914_02_T1/") |> dn_to_reflectance

# Load Shapefile
shp = Shapefile.Table("data/landcover/landcover.shp") |> DataFrame

# Plot Signatures
plot_signatures(landsat, shp, :C_name; colors=cgrad(:tab10))

Utilities

RemoteSensingToolbox.Utils.stack2df — Method

stack2df(rs::RasterStack; dropmissingvals=true)

Convert a RasterStack into a DataFrame.

Parameters

rs: The RasterStack or AbstractSensor from which to extract spectral signatures.

dropmissingvals: Whether or not to drop missing values from the DataFrame. If true, drops points with missing values in at least one layer.

Returns

A DataFrame consisting of rows for each point in the RasterStack and columns for each layer.

Example

julia> landsat = Landsat8("data/LC08_L2SP_043024_20200802_20200914_02_T1/");

julia> stack2df(landsat.stack)
40483825×7 DataFrame
      Row │ B1      B2      B3      B4      B5      B6      B7     
          │ UInt16  UInt16  UInt16  UInt16  UInt16  UInt16  UInt16 
──────────┼────────────────────────────────────────────────────────
        1 │   8348    8345    8798    8216   14454   11201    9070
        2 │   7990    8064    8707    8106   15583   11652    9376
        3 │   8180    8247    8858    8135   17552   12097    9386
        4 │   8400    8362    8952    8262   17836   12772    9797
        5 │   8447    8398    9088    8284   20507   12857    9901
        6 │   8375    8337    8818    8190   16348   12237    9662
        7 │   7742    8005    8915    8257   16683   12465    9862
        8 │   7996    8152    9002    8530   16328   13528   10526
        9 │   8178    8213    8887    8513   17244   13059   10282
    ⋮     │   ⋮       ⋮       ⋮       ⋮       ⋮       ⋮       ⋮
 40483818 │   8609    8835    9707    9924   16742   16214   12380
 40483819 │   8520    8662    9578    9722   16355   15326   11816
 40483820 │   8484    8646    9539    9517   16800   15304   11735
 40483821 │   8645    8854    9775   10064   16739   16707   12667
 40483822 │   8454    8596    9515    9455   17095   15065   11488
 40483823 │   8586    8863    9867   10164   16688   16382   12640
 40483824 │   8601    8823    9684   10050   16210   16211   12439
 40483825 │   8700    8934    9898   10324   16947   16727   12722
                                              40483808 rows omitted