RSToolbox

Documentation for RSToolbox.

Index

RSToolbox.Agriculture

RSToolbox.ColorInfrared

RSToolbox.Geology

RSToolbox.SWIR

RSToolbox.Sensors.AbstractSensor

RSToolbox.Sensors.Landsat7

RSToolbox.Sensors.Landsat8

RSToolbox.Sensors.Sentinel2A

RSToolbox.TrueColor

RSToolbox.Algorithms.mndwi

RSToolbox.Algorithms.nbri

RSToolbox.Algorithms.ndbi

RSToolbox.Algorithms.ndmi

RSToolbox.Algorithms.ndvi

RSToolbox.Algorithms.ndwi

RSToolbox.Algorithms.savi

RSToolbox.Sensors.asraster

RSToolbox.Sensors.blue

RSToolbox.Sensors.dn_to_reflectance

RSToolbox.Sensors.green

RSToolbox.Sensors.nir

RSToolbox.Sensors.red

RSToolbox.Sensors.swir1

RSToolbox.Sensors.swir2

RSToolbox.visualize

Sensors

Sensors are julia structs that wrap a typical Rasters.RasterStack object to provide compatability with many RSToolbox algorithms and methods.

The following methods are supported by all AbstractSensor types:

Base.getindex	return the layer correspinding to the given band name.

Base.length	return the number of layers in the enclosed Rasters.RasterStack.

Base.map	apply a function to each layer in the enclosed Rasters.RasterStack.

Base.write	write layers to file.

Rasters.resample	resample data to a different size and projection, or snap to another object.

Rasters.crop	shrink objects to specific dimension sizes or the extent of another object.

Rasters.extend	extend objects to specific dimension sizes or the extent of another object.

Rasters.trim	trims areas of missing values for arrays and across stack layers.

Rasters.mask	mask an object by a polygon or Raster along X/Y, or other dimensions.

Rasters.replace_missing	replace all missing values in an object and update missingval.

Additionally, asraster can be used to apply a function to the enclosed Rasters.RasterStack.

RSToolbox.Sensors.AbstractSensor — Type

The supertype of all sensor types.

Subtypes should wrap a RasterStack under the field 'stack' and implement the following interface:

blue(X::Sensor)

green(X::Sensor)

red(X::Sensor) 

nir(X::Sensor) 

swir1(X::Sensor)

swir2(X::Sensor)

dn2rs(::Type{<:AbstractSensor})

BandSet(::Type{Landsat8})

Example Implementation

struct Landsat8 <: AbstractSensor
    stack::RasterStack
end

function BandSet(::Type{Landsat8})
    bands = [:B1, :B2, :B3, :B4, :B5, :B6, :B7]
    wavelengths = [440, 480, 560, 655, 865, 1610, 2200]
    return BandSet(bands, wavelengths)
end
    
blue(X::Landsat8) = X[:B2]

green(X::Landsat8) = X[:B3]

red(X::Landsat8) = X[:B4]

nir(X::Landsat8) = X[:B5]

swir1(X::Landsat8) = X[:B6]

swir2(X::Landsat8) = X[:B7]

dn2rs(::Type{Landsat8}) = (scale=0.0000275, offset=-0.2)

RSToolbox.Sensors.Landsat8 — Type

stack::Rasters.RasterStack

Implements the AbstractSensor interface for Landsat 8.

RSToolbox.Sensors.Landsat7 — Type

stack::Rasters.RasterStack

Implements the AbstractSensor interface for Landsat 7.

RSToolbox.Sensors.Sentinel2A — Type

stack::Rasters.RasterStack

Implements the AbstractSensor interface for Sentinel-2A.

RSToolbox.Sensors.red — Function

red(X::AbstractSensor)

Return the red band for the given sensor.

RSToolbox.Sensors.green — Function

green(X::AbstractSensor)

Return the green band for the given sensor.

RSToolbox.Sensors.blue — Function

blue(X::AbstractSensor)

Return the blue band for the given sensor.

RSToolbox.Sensors.nir — Function

nir(X::AbstractSensor)

Return the nir band for the given sensor.

RSToolbox.Sensors.swir1 — Function

swir1(X::AbstractSensor)

Return the swir1 band for the given sensor.

RSToolbox.Sensors.swir2 — Function

swir2(X::AbstractSensor)

Return the swir2 band for the given sensor.

RSToolbox.Sensors.dn_to_reflectance — Function

dn_to_reflectance(X::AbstractSensor)

Transform the raster from Digital Numbers (DN) to reflectance.

RSToolbox.Sensors.asraster — Function

asraster(f, X::AbstractSensor)

Operate on the AbstractSensor as if it was a regular Rasters.RasterStack.

Example

landsat = Landsat8("LC08_L2SP_043024_20200802_20200914_02_T1/")
asraster(landsat) do stack
    map(x -> x .* 0.0001f0, stack)
end

Visualization

RSToolbox.visualize — Function

visualize(r::AbstractRaster, g::AbstractRaster, b::AbstractRaster; lower=0.02, upper=0.98)
visualize(g::AbstractRaster; lower=0.02, upper=0.98)
visualize(img::AbstractSensor, ::Type{TrueColor}; lower=0.02, upper=0.98)
visualize(img::AbstractSensor, ::Type{ColorInfrared}; lower=0.02, upper=0.98)
visualize(img::AbstractSensor, ::Type{SWIR}; lower=0.02, upper=0.98)
visualize(img::AbstractSensor, ::Type{Agriculture}; lower=0.02, upper=0.98)
visualize(img::AbstractSensor, ::Type{Geology}; lower=0.02, upper=0.98)

Visualize a remotely sensed image by applying a histogram stretch. Returns either an RGB or grayscale image compatible with the Images.jl ecosystem.

A number of band combinations are supported for types implementing the AbstractSensor interface.

Example 1

landsat = Landsat8("LC08_L2SP_043024_20200802_20200914_02_T1/")
img = visualize(red(landsat), green(landsat), blue(landsat))
save("truecolor.png", img)

Example 2

landsat = Landsat8("LC08_L2SP_043024_20200802_20200914_02_T1/")
img = visualize(landsat, TrueColor)
save("truecolor.png", img)

RSToolbox.TrueColor — Type

True color band composite.

RSToolbox.ColorInfrared — Type

Color infrared band composite.

RSToolbox.SWIR — Type

SWIR band composite.

RSToolbox.Agriculture — Type

Agriculture band composite.

RSToolbox.Geology — Type

Geology band composite.

Land Cover Indices

RSToolbox.Algorithms.mndwi — Function

mndwi(green::AbstractRaster, swir::AbstractRaster)
mndwi(sensor::AbstractSensor)

Compute the Modified Normalised Difference Water Index (Xu 2006).

MNDWI = (green - swir) / (green + swir)

RSToolbox.Algorithms.ndwi — Function

ndwi(green::AbstractRaster, nir::AbstractRaster)
ndwi(sensor::AbstractSensor)

Compute the Normalized Difference Water Index (McFeeters 1996).

NDWI = (green - nir) / (green + nir)

RSToolbox.Algorithms.ndvi — Function

ndvi(nir::AbstractRaster, red::AbstractRaster)
ndvi(sensor::AbstractSensor)

Compute the Normalized Difference Vegetation Index.

NDVI = (nir - red) / (nir + red)

RSToolbox.Algorithms.savi — Function

savi(nir::AbstractRaster, red::AbstractRaster; L=0.33)
savi(sensor::AbstractSensor; L=0.33)

Compute the Soil Adjusted Vegetation Index (Huete 1988).

SAVI is a vegetation index which attempts to minimize soil brightness influences by introducing a soil-brightness correction factor (L).

L represents the amount of green vegetation cover, which is set to 0.33 by default.

SAVI = ((nir - red) / (nir + red + L)) * (1 + L)

RSToolbox.Algorithms.ndmi — Function

ndmi(nir::AbstractRaster, swir1::AbstractRaster)
ndmi(sensor::AbstractSensor)

Compute the Normalized Difference Moisture Index.

NDMI is sensitive to the moisture levels in vegetation. It is used to monitor droughts and fuel levels in fire-prone areas.

NDMI = (nir - swir1) / (nir + swir1)

RSToolbox.Algorithms.nbri — Function

nbri(nir::AbstractRaster, swir2::AbstractRaster)
nbri(sensor::AbstractSensor)

Compute the Normalized Burn Ratio Index.

NBRI is used to emphasize burned areas.

NBRI = (nir - swir2) / (nir + swir2)

RSToolbox.Algorithms.ndbi — Function

ndbi(swir1::AbstractRaster, nir::AbstractRaster)
ndbi(sensor::AbstractSensor)

Compute the The Normalized Difference Built-up Index

NDBI is used to emphasize urban and built-up areas.

NDBI = (swir1 - nir) / (swir1 + nir)