RemoteSensingToolbox

RemoteSensingToolbox is a pure Julia package intended to provide a collection of tools for visualizing, manipulating, and interpreting remotely sensed imagery.

RemoteSensingToolbox provides a number of utilities for . First, lets load the imagery we want to work with. We're using Landsat 8 imagery in this example, so we'll pass the Landsat8 type to read_bands so it knows how to parse the relevant files from the provided directory. Landsat8 is an instance of AbstractBandset, which is the supertype responsible for allowing many methods within RemoteSensingToolbox to infer sensor-specific information by exploiting Julia's multiple dispatch system.

Bandsets

Bandsets are julia types that encode the sensor-specific information needed for many methods in RemoteSensingToolbox to work without the need for tedious details provided by the end user.

Base.getindex	return the layer correspinding to the given band name.

Base.length	return the number of layers in the enclosed Rasters.RasterStack.

Base.map	apply a function to each layer in the enclosed Rasters.RasterStack.

Base.write	write layers to file.

Rasters.resample	resample data to a different size and projection, or snap to another object.

Rasters.crop	shrink objects to specific dimension sizes or the extent of another object.

Rasters.extend	extend objects to specific dimension sizes or the extent of another object.

Rasters.trim	trims areas of missing values for arrays and across stack layers.

Rasters.mask	mask an object by a polygon or Raster along X/Y, or other dimensions.

Rasters.replace_missing	replace all missing values in an object and update missingval.

Additionally, asraster can be used to apply a function to the enclosed Rasters.RasterStack.

Modules = [RemoteSensingToolbox.Sensors]

Visualization

RemoteSensingToolbox.Agriculture — Type

Agriculture band composite.

RemoteSensingToolbox.ColorInfrared — Type

Color infrared band composite.

RemoteSensingToolbox.Geology — Type

Geology band composite.

RemoteSensingToolbox.SWIR — Type

SWIR band composite.

RemoteSensingToolbox.TrueColor — Type

True color band composite.

RemoteSensingToolbox.visualize — Method

visualize(g::AbstractRaster; lower=0.02, upper=0.98)
visualize(r::AbstractRaster, g::AbstractRaster, b::AbstractRaster; lower=0.02, upper=0.98)
visualize(img::AbstractBandSet, ::Type{TrueColor{AbstractBandset}}; kwargs...)
visualize(img::AbstractBandSet, ::Type{ColorInfrared{AbstractBandset}}; kwargs...)
visualize(img::AbstractBandSet, ::Type{SWIR{AbstractBandset}}; kwargs...)
visualize(img::AbstractBandSet, ::Type{Agriculture{AbstractBandset}}; kwargs...)
visualize(img::AbstractBandSet, ::Type{Geology{AbstractBandset}}; kwargs...)

Visualize a satellite image after applying a histogram stretch. Returns either an RGB or grayscale image compatible with the Images.jl ecosystem.

A number of band combinations are supported for types implementing the AbstractBandSet interface.

Example 1

landsat = read(Landsat8, "LC08_L2SP_043024_20200802_20200914_02_T1/")
img = mndwi(landsat, Landsat8) |> visualize
save("mndwi.png", img)

Example 2

landsat = read(Landsat8, "LC08_L2SP_043024_20200802_20200914_02_T1/")
img = visualize(landsat, TrueColor{Landsat8}; upper=0.90)
save("truecolor.png", img)

Preprocessing

RemoteSensingToolbox.create_tiles — Method

create_tiles(raster, tile::Tuple{Int,Int}; stride=tile)

Slice the given raster into tiles with size tile.

Parameters

raster: The raster to be cut into tiles.

tile: The size of the generated tiles in terms of width x height.

stride: The distance between the top-left corner of each tile. Is equal to tile by default, which produces non-overlapping tiles.

RemoteSensingToolbox.mask_pixels! — Method

mask_pixels!(raster, mask; invert_mask=false)

Drop pixels from a raster according to a given mask. The mask and raster must have the same extent and size.

Parameters

raster: The raster to be masked.

mask: A mask defining which pixels we want to drop. By default, we drop pixels corresponding to mask values of 1.

invert_mask: Treat mask values of 1 as 0 and vice-versa.

RemoteSensingToolbox.mask_pixels — Method

mask_pixels(raster, mask; invert_mask=false)

Drop pixels from a raster according to a given mask. The mask and raster must have the same extent and size.

Parameters

raster: The raster to be masked.

mask: A mask defining which pixels we want to drop. By default, we drop pixels corresponding to mask values of 1.

invert_mask: Treat mask values of 1 as 0 and vice-versa.

RemoteSensingToolbox.tocube — Method

tocube(rs::RasterStack; layers=names(rs))
tocube(rs::AbstractBandset; layers=names(rs))

Transform the multi-layer RasterStack to a multi-band raster.

Parameters

X: The RasterStack or AbstractBandset to be transformed into a multi-band raster.

layers: The layers to include in the new raster.

Example

julia> landsat = Landsat8("LC08_L2SP_043024_20200802_20200914_02_T1");
julia> tocube(landsat)
7821×7921×7 Raster{Float32,3} B1 with dimensions: 
  X Projected{Float64} LinRange{Float64}(493785.0, 728385.0, 7821) ForwardOrdered Regular Points crs: WellKnownText,
  Y Projected{Float64} LinRange{Float64}(5.84638e6, 5.60878e6, 7921) ReverseOrdered Regular Points crs: WellKnownText,
  Band Categorical{Int64} 1:7 ForwardOrdered
extent: Extent(X = (493785.0, 728385.0), Y = (5.608785e6, 5.846385e6), Band = (1, 7))
missingval: 0.0f0
crs: PROJCS["WGS 84 / UTM zone 11N",GEOGCS["WGS 84",DATUM["WGS_1984",SPHEROID["WGS 84",6378137,298.257223563,AUTHORITY["EPSG","7030"]],AUTHORITY["EPSG","6326"]],PRIMEM["Greenwich",0,AUTHORITY["EPSG","8901"]],UNIT["degree",0.0174532925199433,AUTHORITY["EPSG","9122"]],AUTHORITY["EPSG","4326"]],PROJECTION["Transverse_Mercator"],PARAMETER["latitude_of_origin",0],PARAMETER["central_meridian",-117],PARAMETER["scale_factor",0.9996],PARAMETER["false_easting",500000],PARAMETER["false_northing",0],UNIT["metre",1,AUTHORITY["EPSG","9001"]],AXIS["Easting",EAST],AXIS["Northing",NORTH],AUTHORITY["EPSG","32611"]]
parent:
[:, :, 1]
           5.84638e6  5.84636e6  5.84632e6  5.8463e6  5.84626e6  …  5.60894e6  5.6089e6  5.60888e6  5.60884e6  5.60882e6  5.60878e6
 493785.0  0.0        0.0        0.0        0.0       0.0           0.0        0.0       0.0        0.0        0.0        0.0
      ⋮                                               ⋮          ⋱                                             ⋮          
 728355.0  0.0        0.0        0.0        0.0       0.0           0.0        0.0       0.0        0.0        0.0        0.0
 728385.0  0.0        0.0        0.0        0.0       0.0           0.0        0.0       0.0        0.0        0.0        0.0
[and 6 more slices...]

Land Cover Indices

RemoteSensingToolbox.mndwi — Method

mndwi(green::AbstractRaster, swir::AbstractRaster)
mndwi(stack::AbstractRasterStack, ::Type{AbstractBandset})

Compute the Modified Normalised Difference Water Index (Xu 2006).

MNDWI = (green - swir) / (green + swir)

RemoteSensingToolbox.nbri — Method

nbri(nir::AbstractRaster, swir2::AbstractRaster)
nbri(stack::AbstractRasterStack, ::Type{AbstractBandset})

Compute the Normalized Burn Ratio Index.

NBRI is used to emphasize burned areas.

NBRI = (nir - swir2) / (nir + swir2)

RemoteSensingToolbox.ndbi — Method

ndbi(swir1::AbstractRaster, nir::AbstractRaster)
ndbi(stack::AbstractRasterStack, ::Type{AbstractBandset})

Compute the The Normalized Difference Built-up Index

NDBI is used to emphasize urban and built-up areas.

NDBI = (swir1 - nir) / (swir1 + nir)

RemoteSensingToolbox.ndmi — Method

ndmi(nir::AbstractRaster, swir1::AbstractRaster)
ndmi(stack::AbstractRasterStack, ::Type{AbstractBandset})

Compute the Normalized Difference Moisture Index.

NDMI is sensitive to the moisture levels in vegetation. It is used to monitor droughts and fuel levels in fire-prone areas.

NDMI = (nir - swir1) / (nir + swir1)

RemoteSensingToolbox.ndvi — Method

ndvi(nir::AbstractRaster, red::AbstractRaster)
ndvi(stack::AbstractRasterStack, ::Type{AbstractBandset})

Compute the Normalized Difference Vegetation Index.

NDVI = (nir - red) / (nir + red)

RemoteSensingToolbox.ndwi — Method

ndwi(green::AbstractRaster, nir::AbstractRaster)
ndwi(stack::AbstractRasterStack, ::Type{AbstractBandset})

Compute the Normalized Difference Water Index (McFeeters 1996).

NDWI = (green - nir) / (green + nir)

RemoteSensingToolbox.savi — Method

savi(nir::AbstractRaster, red::AbstractRaster; L=0.33)
savi(stack::AbstractRasterStack, ::Type{AbstractBandset}; L=0.33)

Compute the Soil Adjusted Vegetation Index (Huete 1988).

SAVI is a vegetation index which attempts to minimize soil brightness influences by introducing a soil-brightness correction factor (L).

L represents the amount of green vegetation cover, which is set to 0.33 by default.

SAVI = ((nir - red) / (nir + red + L)) * (1 + L)

Spectral Analysis

Modules = [RemoteSensingToolbox.Spectral]

Transformations

Modules = [RemoteSensingToolbox.Transformations]