RemoteSensingToolbox

Documentation for RemoteSensingToolbox.

Sensors

Sensors are julia structs that wrap a typical Rasters.RasterStack object to provide compatability with many RemoteSensingToolbox algorithms and methods.

The following methods are supported by all AbstractSensor types:

Base.getindex	return the layer correspinding to the given band name.

Base.length	return the number of layers in the enclosed Rasters.RasterStack.

Base.map	apply a function to each layer in the enclosed Rasters.RasterStack.

Base.write	write layers to file.

Rasters.resample	resample data to a different size and projection, or snap to another object.

Rasters.crop	shrink objects to specific dimension sizes or the extent of another object.

Rasters.extend	extend objects to specific dimension sizes or the extent of another object.

Rasters.trim	trims areas of missing values for arrays and across stack layers.

Rasters.mask	mask an object by a polygon or Raster along X/Y, or other dimensions.

Rasters.replace_missing	replace all missing values in an object and update missingval.

Additionally, asraster can be used to apply a function to the enclosed Rasters.RasterStack.

RemoteSensingToolbox.Sensors.AbstractSensor — Type

The supertype of all sensor types.

Subtypes should wrap a RasterStack under the field 'stack' and implement the following interface:

blue(X::Sensor)

green(X::Sensor)

red(X::Sensor) 

nir(X::Sensor) 

swir1(X::Sensor)

swir2(X::Sensor)

dn2rs(::Type{<:AbstractSensor})

BandSet(::Type{Landsat8})

Example Implementation

struct Landsat8 <: AbstractSensor
    stack::RasterStack
end

function BandSet(::Type{Landsat8})
    bands = [:B1, :B2, :B3, :B4, :B5, :B6, :B7]
    wavelengths = [440, 480, 560, 655, 865, 1610, 2200]
    return BandSet(bands, wavelengths)
end
    
blue(X::Landsat8) = X[:B2]

green(X::Landsat8) = X[:B3]

red(X::Landsat8) = X[:B4]

nir(X::Landsat8) = X[:B5]

swir1(X::Landsat8) = X[:B6]

swir2(X::Landsat8) = X[:B7]

dn2rs(::Type{Landsat8}) = (scale=0.0000275, offset=-0.2)

RemoteSensingToolbox.Sensors.BandSet — Type

A struct for storing the band names and associated wavelengths of a particular sensor.

It is expected that instances of AbstractSensor implement a BandSet constructor.

The central wavelength for a given band can be recovered by calling the BandSet.

Example

julia> bandset = BandSet(Sentinel2A);
julia> bandset(:B8A)
842.0

RemoteSensingToolbox.Sensors.DESIS — Type

stack::Rasters.RasterStack

Implements the AbstractSensor interface for DESIS.

RemoteSensingToolbox.Sensors.Landsat7 — Type

stack::Rasters.RasterStack

Implements the AbstractSensor interface for Landsat 7.

RemoteSensingToolbox.Sensors.Landsat8 — Type

stack::Rasters.RasterStack

Implements the AbstractSensor interface for Landsat 8.

RemoteSensingToolbox.Sensors.Sentinel2A — Type

stack::Rasters.RasterStack

Implements the AbstractSensor interface for Sentinel-2A.

RemoteSensingToolbox.Sensors.asraster — Method

asraster(f, X::AbstractSensor)

Operate on the AbstractSensor as if it was a regular Rasters.RasterStack.

Example

landsat = Landsat8("LC08_L2SP_043024_20200802_20200914_02_T1/")
asraster(landsat) do stack
    map(x -> x .* 0.0001f0, stack)
end

RemoteSensingToolbox.Sensors.blue — Method

blue(X::AbstractSensor)

Return the blue band for the given sensor.

RemoteSensingToolbox.Sensors.dn2rs — Method

dn2rs(::Type{<:AbstractSensor})

Return the scale and offset required to convert DN to reflectance for the given sensor type.

Example

julia> dn2rs(Landsat8)
(scale = 2.75e-5, offset = -0.2)

RemoteSensingToolbox.Sensors.dn_to_reflectance — Method

dn_to_reflectance(X::AbstractSensor)

Transform the raster from Digital Numbers (DN) to reflectance.

RemoteSensingToolbox.Sensors.green — Method

green(X::AbstractSensor)

Return the green band for the given sensor.

RemoteSensingToolbox.Sensors.nir — Method

nir(X::AbstractSensor)

Return the nir band for the given sensor.

RemoteSensingToolbox.Sensors.red — Method

red(X::AbstractSensor)

Return the red band for the given sensor.

RemoteSensingToolbox.Sensors.swir1 — Method

swir1(X::AbstractSensor)

Return the swir1 band for the given sensor.

RemoteSensingToolbox.Sensors.swir2 — Method

swir2(X::AbstractSensor)

Return the swir2 band for the given sensor.

Visualization

RemoteSensingToolbox.Agriculture — Type

Agriculture band composite.

RemoteSensingToolbox.ColorInfrared — Type

Color infrared band composite.

RemoteSensingToolbox.Geology — Type

Geology band composite.

RemoteSensingToolbox.SWIR — Type

SWIR band composite.

RemoteSensingToolbox.TrueColor — Type

True color band composite.

RemoteSensingToolbox.visualize — Method

visualize(r::AbstractRaster, g::AbstractRaster, b::AbstractRaster; lower=0.02, upper=0.98)
visualize(g::AbstractRaster; lower=0.02, upper=0.98)
visualize(img::AbstractSensor, ::Type{TrueColor}; lower=0.02, upper=0.98)
visualize(img::AbstractSensor, ::Type{ColorInfrared}; lower=0.02, upper=0.98)
visualize(img::AbstractSensor, ::Type{SWIR}; lower=0.02, upper=0.98)
visualize(img::AbstractSensor, ::Type{Agriculture}; lower=0.02, upper=0.98)
visualize(img::AbstractSensor, ::Type{Geology}; lower=0.02, upper=0.98)

Visualize a remotely sensed image by applying a histogram stretch. Returns either an RGB or grayscale image compatible with the Images.jl ecosystem.

A number of band combinations are supported for types implementing the AbstractSensor interface.

Example 1

landsat = Landsat8("LC08_L2SP_043024_20200802_20200914_02_T1/")
img = visualize(red(landsat), green(landsat), blue(landsat))
save("truecolor.png", img)

Example 2

landsat = Landsat8("LC08_L2SP_043024_20200802_20200914_02_T1/")
img = visualize(landsat, TrueColor)
save("truecolor.png", img)

Land Cover Indices

RemoteSensingToolbox.Algorithms.mndwi — Method

mndwi(green::AbstractRaster, swir::AbstractRaster)
mndwi(sensor::AbstractSensor)

Compute the Modified Normalised Difference Water Index (Xu 2006).

MNDWI = (green - swir) / (green + swir)

RemoteSensingToolbox.Algorithms.nbri — Method

nbri(nir::AbstractRaster, swir2::AbstractRaster)
nbri(sensor::AbstractSensor)

Compute the Normalized Burn Ratio Index.

NBRI is used to emphasize burned areas.

NBRI = (nir - swir2) / (nir + swir2)

RemoteSensingToolbox.Algorithms.ndbi — Method

ndbi(swir1::AbstractRaster, nir::AbstractRaster)
ndbi(sensor::AbstractSensor)

Compute the The Normalized Difference Built-up Index

NDBI is used to emphasize urban and built-up areas.

NDBI = (swir1 - nir) / (swir1 + nir)

RemoteSensingToolbox.Algorithms.ndmi — Method

ndmi(nir::AbstractRaster, swir1::AbstractRaster)
ndmi(sensor::AbstractSensor)

Compute the Normalized Difference Moisture Index.

NDMI is sensitive to the moisture levels in vegetation. It is used to monitor droughts and fuel levels in fire-prone areas.

NDMI = (nir - swir1) / (nir + swir1)

RemoteSensingToolbox.Algorithms.ndvi — Method

ndvi(nir::AbstractRaster, red::AbstractRaster)
ndvi(sensor::AbstractSensor)

Compute the Normalized Difference Vegetation Index.

NDVI = (nir - red) / (nir + red)

RemoteSensingToolbox.Algorithms.ndwi — Method

ndwi(green::AbstractRaster, nir::AbstractRaster)
ndwi(sensor::AbstractSensor)

Compute the Normalized Difference Water Index (McFeeters 1996).

NDWI = (green - nir) / (green + nir)

RemoteSensingToolbox.Algorithms.savi — Method

savi(nir::AbstractRaster, red::AbstractRaster; L=0.33)
savi(sensor::AbstractSensor; L=0.33)

Compute the Soil Adjusted Vegetation Index (Huete 1988).

SAVI is a vegetation index which attempts to minimize soil brightness influences by introducing a soil-brightness correction factor (L).

L represents the amount of green vegetation cover, which is set to 0.33 by default.

SAVI = ((nir - red) / (nir + red + L)) * (1 + L)

Spectral Analysis

RemoteSensingToolbox.Spectral.extract_signatures — Method

extract_signatures(rs::AbstractSensor, shp::DataFrame, label::Symbol)
extract_signatures(rs::RasterStack, shp::DataFrame, label::Symbol)

Extract signatures from the given RasterStack or AbstractSensor within regions specified by a given shapefile.

Parameters

rs: The RasterStack or AbstractSensor from which to extract spectral signatures.

shp: A shapefile stored as a DataFrame with a :geometry column storing a GeoInterface.jl compatible geometry and a label column indicating the land cover type.

label: The column in shp in which the land cover class is stored.

Returns

A DataFrame consisting of rows for each extracted signature and columns storing the respective bands and land cover type.

Example

julia> landsat = Landsat8("data/LC08_L2SP_043024_20200802_20200914_02_T1/") |> dn_to_reflectance;

julia> shp = Shapefile.Table("data/landcover/landcover.shp") |> DataFrame;

julia> extract_signatures(landsat, shp, :cover)
3195×8 DataFrame
  Row │ B1          B2         B3         B4         B5        B6         B7         label      
      │ Float32     Float32    Float32    Float32    Float32   Float32    Float32    String     
──────┼─────────────────────────────────────────────────────────────────────────────────────────
    1 │  0.05102    0.0891075  0.156317   0.198558   0.482055  0.267197   0.139103   hail scar
    2 │  0.054815   0.09238    0.16333    0.207742   0.481203  0.270415   0.140973   hail scar
    3 │  0.0561625  0.0942225  0.16564    0.208787   0.477875  0.273082   0.14353    hail scar
    4 │  0.057015   0.0941125  0.16146    0.20425    0.47991   0.271267   0.141715   hail scar
    5 │  0.044695   0.0839375  0.152852   0.193525   0.465198  0.258012   0.133217   hail scar
  ⋮   │     ⋮           ⋮          ⋮          ⋮         ⋮          ⋮          ⋮          ⋮
 3192 │ -0.0024675  0.011475   0.0757425  0.0416975  0.540658  0.0856975  0.035785   vegetation
 3193 │ -0.001945   0.0115575  0.0766225  0.0419725  0.535488  0.08567    0.036005   vegetation
 3194 │ -0.0023025  0.012135   0.0773925  0.0427975  0.523745  0.08556    0.0364175  vegetation
 3195 │ -0.0019725  0.0119425  0.0767875  0.04211    0.523745  0.085065   0.0360325  vegetation
                                                                               3186 rows omitted

RemoteSensingToolbox.Spectral.plot_signatures — Method

plot_signatures(rs::AbstractSensor, shp::DataFrame, label::Symbol, dst::String)
plot_signatures(rs::RasterStack, shp::DataFrame, bandset::BandSet, label::Symbol, dst::String)

Plot spectral signatures for each land cover type specified in a given shapefile.

Parameters

rs: The RasterStack or AbstractSensor from which to extract spectral signatures.

shp: A shapefile stored as a DataFrame with a :geometry column storing a GeoInterface.jl compatible geometry and a label column indicating the land cover type.

bandset: The BandSet for the provided sensor specifying the available bands and associated wavelengths in nm. Inferred for AbstractSensor.

label: The column in shp in which the land cover class is stored.

dst: The destination at which to save the plot. Supports all file extensions supported by CairoMakie.

Example

# Read Landsat And Convert DNs To Reflectance
landsat = Landsat8("data/LC08_L2SP_043024_20200802_20200914_02_T1/") |> dn_to_reflectance

# Load Shapefile
shp = Shapefile.Table("data/landcover/landcover.shp") |> DataFrame

# Plot Signatures
plot_signatures(landsat, shp, :cover, "landsat_sigs.png")