RSToolbox

Documentation for RSToolbox.

RSToolbox.align_rasters

RSToolbox.efficient_read

RSToolbox.extract_raster_data

RSToolbox.linear_stretch

RSToolbox.raster_to_image

RSToolbox.align_rasters — Method

Resample a collection of rasters to a common extent and resolution

RSToolbox.efficient_read — Method

Read a raster from disk into memory. Return immediately if raster has already been read.

RSToolbox.extract_raster_data — Method

Extract the raw raster data, discarding all dimensional information.

RSToolbox.linear_stretch — Method

Adjust image histogram by performing a linear stretch to squeeze all values between the percentiles lower and upper into the range [0,1].

RSToolbox.raster_to_image — Method

Turn a raster into an image compatible with Images.jl.