RemoteSensingToolbox

RemoteSensingToolbox is a pure Julia package built on top of Rasters.jl for visualizing, analyzing, and manipulating remotely sensed imagery. Most methods expect either an AbstractRaster or AbstractRasterStack as input and return the same. The most important exception to this rule is visualize, which returns an Array of either Gray or RGB pixels, depending on whether the visualization is intended to be in color or grayscale. The result is that the output of visualize will be automatically displayed inside Pluto.

Features

RemoteSensingToolbox is a work in progress. This means that new features are being added and existing features are subject to change. To contribute to this project, please create an issue on GitHub or open a pull request. A summary of both existing and future features are provided below:

Feature	Description	Implemented

Visualization	Visualize images with various band composites	Yes

Land Cover Indices	Calculate indices such as MNDWI and NDVI	Yes

QA and SCL Decoding	Decode Quality Assurance and Scene Classification masks	Yes

Pixel Masking	Mask pixels to remove objects such as clouds or shadows	Yes

PCA	Perform PCA analysis, transformation, and reconstruction	Yes

MNF	Minimum Noise Fraction transformation and reconstruction	Yes

Signature Analysis	Visualize spectral signatures for different land cover types	Yes

Land Cover Classification	Exposes an MLJ interface for classifying land cover types	No

Endmember Extraction	Extract spectral endmembers from an image	No

Spectral Unmixing	Perform spectral unmixing under a given endmember library	No

Rasters.jl

RemoteSensingToolbox is intended to be used in conjunction with the wider Julia ecosystem and as such, seeks to avoid duplicating functinalities provided by other packages. As the majority of methods accept and return AbstractRaster or AbstractRasterStack objects, users should be able to call methods from Rasters.jl at any point in the processing pipeline. A summary of common functionalities offered by Rasters is provided below:

Method	Description

mosaic	Join rasters covering different extents into a single array or file.

crop	Shrink objects to specific dimension sizes or the extent of another object.

extend	Extend objects to specific dimension sizes or the extent of another object.

trim	Trims areas of missing values for arrays and across stack layers.

resample	Resample data to a different size and projection, or snap to another object.

mask	Mask a raster by a polygon or the non-missing values of another Raster.

replace_missing	Replace all missing values in a raster and update missingval.

extract	Extract raster values from points or geometries.

zonal	Calculate zonal statistics for a raster masked by geometries.

Bandsets

Bandsets are Julia types that encode the sensor-specific information needed for many methods in RemoteSensingToolbox to work without the need for tedious details from the end user. Such methods typically expect the sensor type to be passed in as the first argument, which then triggers Julia's multiple dispatch to execute the appropriate sensor-specific procedures. Bandsets are provided for several common sensors including Sentinel-2, Landsat 7, and Landsat 8.

Method	Description

bands	Return the band names in order from shortest to longest wavelength.

wavelengths	Return the central wavelengths for all bands from shortest to longest.

wavelength	Return the central wavelength for the specified band.

blue	Return the blue band for the given sensor.

green	Return the green band for the given sensor.

red	Return the red band for the given sensor.

nir	Return the nir band for the given sensor.

swir1	Return the swir1 band for the given sensor.

swir2	Return the swir2 band for the given sensor.

read_bands	Read bands from the specified directory into a Raster or RasterStack.

read_qa	Reads the QA or scene classification file from the provided file or directory.

dn_to_reflectance	Decodes digital numbers to reflectance.

RemoteSensingToolbox.Bandsets.AbstractBandset — Type

The supertype of all sensor types. Provides sensor-specific information to many RemoteSensingToolbox methods.

RemoteSensingToolbox.Bandsets.DESIS — Type

Implements the AbstractBandset interface for Landsat 7.

RemoteSensingToolbox.Bandsets.Landsat8 — Type

Implements the AbstractBandset interface for Landsat 8.

RemoteSensingToolbox.Bandsets.Landsat7 — Type

Implements the AbstractBandset interface for Landsat 7.

RemoteSensingToolbox.Bandsets.Sentinel2 — Type

Implements the AbstractBandset interface for Sentinel 2.

RemoteSensingToolbox.Bandsets.red — Function

red(::Type{AbstractBandset})
red(raster, ::Type{AbstractBandset})

Return the red band for the given sensor.

RemoteSensingToolbox.Bandsets.green — Function

green(::Type{AbstractBandset})
green(raster, ::Type{AbstractBandset})

Return the green band for the given sensor.

RemoteSensingToolbox.Bandsets.blue — Function

blue(::Type{AbstractBandset})
blue(raster, ::Type{AbstractBandset})

Return the blue band for the given sensor.

RemoteSensingToolbox.Bandsets.nir — Function

nir(::Type{AbstractBandset})
nir(raster, ::Type{AbstractBandset})

Return the nir band for the given sensor.

RemoteSensingToolbox.Bandsets.swir1 — Function

swir1(::Type{AbstractBandset})
swir1(raster, ::Type{AbstractBandset})

Return the swir1 band for the given sensor.

RemoteSensingToolbox.Bandsets.swir2 — Function

swir2(::Type{AbstractBandset})
swir2(raster, ::Type{AbstractBandset})

Return the swir2 band for the given sensor.

RemoteSensingToolbox.Bandsets.bands — Function

bands(::Type{AbstractBandset})

Return the band names in order from shortest to longest wavelength.

RemoteSensingToolbox.Bandsets.wavelengths — Function

wavelengths(::Type{AbstractBandset})

Return the central wavelengths for all bands in order from shortest to longest.

RemoteSensingToolbox.Bandsets.wavelength — Function

wavelength(::Type{AbstractBandset}, band::Symbol)

Return the central wavelength for the corresponding band.

RemoteSensingToolbox.Bandsets.parse_band — Function

parse_band(::Type{AbstractBandset}, filename::String)

Parses the band name from the given file path.

Returns either the band as a Symbol or nothing if no band could be parsed.

If the file is a multi-band raster, returns the names of all bands as a Vector{Symbol}.

RemoteSensingToolbox.Bandsets.read_bands — Function

function read_bands(bandset::Type{AbstractBandset}, dir::String)

Read the bands from the given directory into a RasterStack.

Parameters

bandset: The sensor type to which the bands belong.

dir: The directory in which the bands can be read.

Example

julia> landsat = read_bands(Landsat8, "data/LC08_L2SP_043024_20200802_20200914_02_T1/")
RasterStack with dimensions: 
  X Projected{Float64} LinRange{Float64}(493785.0, 728385.0, 7821) ForwardOrdered Regular Points crs: WellKnownText,
  Y Projected{Float64} LinRange{Float64}(5.84638e6, 5.60878e6, 7921) ReverseOrdered Regular Points crs: WellKnownText
and 7 layers:
  :B1 UInt16 dims: X, Y (7821×7921)
  :B2 UInt16 dims: X, Y (7821×7921)
  :B3 UInt16 dims: X, Y (7821×7921)
  :B4 UInt16 dims: X, Y (7821×7921)
  :B5 UInt16 dims: X, Y (7821×7921)
  :B6 UInt16 dims: X, Y (7821×7921)
  :B7 UInt16 dims: X, Y (7821×7921)

RemoteSensingToolbox.Bandsets.read_qa — Function

read_qa(bandset::Type{AbstractBandSet}, src::String)

Read and decode the quality assurance mask for the given AbstractBandset.

Parameters

bandset: A subtype of AbstractBandset.

src: Either a directory containing the quality assurance mask named according to standard conventions or the file itself.

Returns

The decoded quality assurance mask as a RasterStack. Encodes masked values as 1 and non-masked values as 0.

Example

julia> qa = read_qa(Landsat8, "data/LC08_L2SP_043024_20200802_20200914_02_T1/")
RasterStack with dimensions: 
  X Projected{Float64} LinRange{Float64}(493785.0, 728385.0, 7821) ForwardOrdered Regular Points crs: WellKnownText,
  Y Projected{Float64} LinRange{Float64}(5.84638e6, 5.60878e6, 7921) ReverseOrdered Regular Points crs: WellKnownText
and 7 layers:
  :dilated_cloud UInt8 dims: X, Y (7821×7921)
  :cirrus        UInt8 dims: X, Y (7821×7921)
  :cloud         UInt8 dims: X, Y (7821×7921)
  :cloud_shadow  UInt8 dims: X, Y (7821×7921)
  :snow          UInt8 dims: X, Y (7821×7921)
  :clear         UInt8 dims: X, Y (7821×7921)
  :water         UInt8 dims: X, Y (7821×7921)

RemoteSensingToolbox.Bandsets.dn_to_reflectance — Function

dn_to_reflectance(bandset::Type{AbstractBandset}, raster; clamp_values=false)

Transform the raster from Digital Numbers (DN) to reflectance.

Parameters

bandset: A subtype of AbstractBandset.

raster: The AbstractRasterStack or AbstractRaster to be converted to reflectance.

clamp_values: Indicates whether to clamp reflectances into the range (0.0, 1.0] (default = false).

Example

landsat = read(Landsat8, "LC08_L2SP_043024_20200802_20200914_02_T1/")
landsat_sr = dn_to_reflectance(Landsat8, landsat)

Visualization

Remotely sensed imagery is typically encoded as either UInt16 or Int16 values. However, many products only actually use the first 12 bits for storing information. The result is that naive visualization methods will produce a near-black image, since the maximum possible brightness will be located in the lower range of values provided by the 16 bit encoding. To address this, we need to perform a linear stretch before visualizing an image. Additionally, many satellites have more than three bands, which motivates the use of band combinations to emphasize certain features and land cover types.

Method/Type	Description

TrueColor	The true color band combination produces RGB images that are familiar to the human eye.

Agriculture	Used for crop monitoring and emphasizing healthy vegetation.

Geology	Emphasizes geological formations, lithology features, and faults.

ColorInfrared	Highlights vegetation in red, water in blue, and urban areas in grey.

SWIR	Emphasizes dense vegetation in dark green and sparse vegetation in lighter shades.

visualize	Visualize a Raster or RasterStack containing one or more satellite bands.

RemoteSensingToolbox.Agriculture — Type

Agriculture band composite.

RemoteSensingToolbox.ColorInfrared — Type

Color infrared band composite.

RemoteSensingToolbox.Geology — Type

Geology band composite.

RemoteSensingToolbox.SWIR — Type

SWIR band composite.

RemoteSensingToolbox.TrueColor — Type

True color band composite.

RemoteSensingToolbox.visualize — Method

visualize(g::AbstractRaster; lower=0.02, upper=0.98)
visualize(r::AbstractRaster, g::AbstractRaster, b::AbstractRaster; lower=0.02, upper=0.98)
visualize(img::AbstractRasterStack, ::Type{TrueColor{AbstractBandset}}; kwargs...)
visualize(img::AbstractRasterStack, ::Type{ColorInfrared{AbstractBandset}}; kwargs...)
visualize(img::AbstractRasterStack, ::Type{SWIR{AbstractBandset}}; kwargs...)
visualize(img::AbstractRasterStack, ::Type{Agriculture{AbstractBandset}}; kwargs...)
visualize(img::AbstractRasterStack, ::Type{Geology{AbstractBandset}}; kwargs...)

Visualize a satellite image after applying a histogram stretch. Returns either an RGB or grayscale image compatible with the Images.jl ecosystem.

A number of band combinations are supported for types implementing the AbstractBandSet interface.

Example 1

landsat = read_bands(Landsat8, "LC08_L2SP_043024_20200802_20200914_02_T1/")
mndwi(landsat, Landsat8) |> visualize

Example 2

landsat = read_bands(Landsat8, "LC08_L2SP_043024_20200802_20200914_02_T1/")
visualize(landsat, TrueColor{Landsat8}; upper=0.90)

Land Cover Indices

Land cover indices are used to highlight different types of land cover. For example, the Modified Normalized Difference Water Index (MNDWI) is used to highlight water while diminishing built-up areas. Each index is expressed as a function of two or more bands. RemoteSensingToolbox can automatically select the appropriate bands for a given index by providing the Bandset to which the image belongs. Lower-level variants are also provided for manual band selection.

Method	Description

mndwi	The Modified Normalized Difference Index is used to highlight surface water features.

nbri	The Normalized Burn Ratio Index is used to emphasize burned areas.

ndbi	The Normalized Difference Built-Up Index is used to empasize urban areas.

ndmi	The Normalized Difference Moisture Index is used to monitor droughts and dry areas.

ndvi	The Normalized Difference Vegetation Index is used to emphasize vegetation.

ndwi	The Normalized Difference Water Index is used to highlight water bodies and vegetation.

savi	The Soil Adjusted Vegetation Index is an adjusted variant of NDVI.

RemoteSensingToolbox.mndwi — Method

mndwi(green::AbstractRaster, swir::AbstractRaster)
mndwi(stack::AbstractRasterStack, ::Type{AbstractBandset})

Compute the Modified Normalised Difference Water Index (Xu 2006).

MNDWI = (green - swir) / (green + swir)

RemoteSensingToolbox.nbri — Method

nbri(nir::AbstractRaster, swir2::AbstractRaster)
nbri(stack::AbstractRasterStack, ::Type{AbstractBandset})

Compute the Normalized Burn Ratio Index.

NBRI is used to emphasize burned areas.

NBRI = (nir - swir2) / (nir + swir2)

RemoteSensingToolbox.ndbi — Method

ndbi(swir1::AbstractRaster, nir::AbstractRaster)
ndbi(stack::AbstractRasterStack, ::Type{AbstractBandset})

Compute the The Normalized Difference Built-up Index

NDBI is used to emphasize urban and built-up areas.

NDBI = (swir1 - nir) / (swir1 + nir)

RemoteSensingToolbox.ndmi — Method

ndmi(nir::AbstractRaster, swir1::AbstractRaster)
ndmi(stack::AbstractRasterStack, ::Type{AbstractBandset})

Compute the Normalized Difference Moisture Index.

NDMI is sensitive to the moisture levels in vegetation. It is used to monitor droughts and fuel levels in fire-prone areas.

NDMI = (nir - swir1) / (nir + swir1)

RemoteSensingToolbox.ndvi — Method

ndvi(nir::AbstractRaster, red::AbstractRaster)
ndvi(stack::AbstractRasterStack, ::Type{AbstractBandset})

Compute the Normalized Difference Vegetation Index.

NDVI = (nir - red) / (nir + red)

RemoteSensingToolbox.ndwi — Method

ndwi(green::AbstractRaster, nir::AbstractRaster)
ndwi(stack::AbstractRasterStack, ::Type{AbstractBandset})

Compute the Normalized Difference Water Index (McFeeters 1996).

NDWI = (green - nir) / (green + nir)

RemoteSensingToolbox.savi — Method

savi(nir::AbstractRaster, red::AbstractRaster; L=0.33)
savi(stack::AbstractRasterStack, ::Type{AbstractBandset}; L=0.33)

Compute the Soil Adjusted Vegetation Index (Huete 1988).

SAVI is a vegetation index which attempts to minimize soil brightness influences by introducing a soil-brightness correction factor (L).

L represents the amount of green vegetation cover, which is set to 0.33 by default.

SAVI = ((nir - red) / (nir + red + L)) * (1 + L)

Spectral Analysis

Spectral analysis involves studying the relationships between different materials and their corresponding spectral signatures. Due to the interactions between light and matter, each signature is unique to the material that emitted it. We can exploit this fact to assign a label to each pixel, or even estimate the abundances of different materials at a sub-pixel level.

Method	Description

extract_signatures	Extract spectral signatures and their corresponding land cover type.

plot_signatures	Plot the spectral signatures for each type of land cover specified in a shapefile.

plot_signatures!	The mutating version of plot_signatures. Writes to a Makie.Axis object.

RemoteSensingToolbox.Spectral.extract_signatures — Method

extract_signatures(stack::AbstractRasterStack, shp, label::Symbol; drop_missing=true)

Extract signatures from the given RasterStack within regions specified by a provided shapefile.

Parameters

stack: The RasterStack from which to extract spectral signatures.

shp: A Tables.jl compatible object containing a :geometry column storing a GeoInterface.jl compatible geometry and a label column indicating the land cover type.

label: The column in shp corresponding to the land cover type.

drop_missing: Drop all rows with at least one missing value in either the bands or labels (default = true).

Returns

A RasterTable consisting of rows for each observed signature and columns storing the respective bands and land cover type.

Example

julia> landsat = @pipe read_bands(Landsat8, "data/LC08_L2SP_043024_20200802_20200914_02_T1/") |> dn_to_reflectance(Landsat8, _)

julia> shp = Shapefile.Table("data/landcover/landcover.shp") |> DataFrame

julia> extract_signatures(landsat, shp, :C_name) |> DataFrame
1925×8 DataFrame
  Row │ B1         B2         B3        B4        B5        B6        B7        label     
      │ Float32    Float32    Float32   Float32   Float32   Float32   Float32   String    
──────┼───────────────────────────────────────────────────────────────────────────────────
    1 │ 0.057235   0.10547    0.188932  0.24847   0.513405  0.315323  0.166107  Hail Scar
    2 │ 0.0574     0.105415   0.188712  0.24946   0.513735  0.317908  0.167593  Hail Scar
    3 │ 0.0584175  0.107175   0.19182   0.253887  0.515577  0.319585  0.168087  Hail Scar
    4 │ 0.0583625  0.10723    0.190995  0.253393  0.514615  0.317715  0.167097  Hail Scar
    5 │ 0.05806    0.10745    0.189895  0.250368  0.50774   0.315048  0.165475  Hail Scar
  ⋮   │     ⋮          ⋮         ⋮         ⋮         ⋮         ⋮         ⋮          ⋮
 1921 │ 0.077145   0.112097   0.17081   0.19677   0.25705   0.320877  0.28015   Built Up
 1922 │ 0.10712    0.136765   0.19666   0.225068  0.268655  0.33284   0.302288  Built Up
 1923 │ 0.0856425  0.123235   0.185853  0.211675  0.273     0.373623  0.351072  Built Up
 1924 │ 0.088145   0.119578   0.17059   0.204442  0.27487   0.355857  0.325553  Built Up
 1925 │ 0.0518725  0.0816275  0.148673  0.15189   0.28774   0.309218  0.29786   Built Up
                                                                         1915 rows omitted

RemoteSensingToolbox.Spectral.plot_signatures! — Method

plot_signatures!(ax, bandset::Type{<:AbstractBandset}, raster::AbstractRasterStack, shp, label::Symbol; colors=wong_colors())

Plot spectral signatures for each land cover type specified in a given shapefile by mutating a Makie.Axis object.

Accepts the same keywords as Makie.lines!.

Parameters

ax: The Makie.Axis into which we want to draw our plot.

bandset: The sensor type to which the signatures belong.

raster: A RasterStack from which we want to extract the spectral signatures.

shp: A Tables.jl compatible object containing a :geometry column storing a GeoInterface.jl compatible geometry and a label column indicating the land cover type.

label: The column in shp corresponding to the land cover type.

colors: The color scheme used by the plot.

Example

# Read Images And Convert DNs To Reflectance
landsat = @pipe read_bands(Landsat8, "data/LC08_L2SP_043024_20200802_20200914_02_T1/") |> dn_to_reflectance(Landsat8, _)
sentinel = @pipe read_bands(Sentinel2, "data/T11UPT_20200804T183919/R60m/") |> dn_to_reflectance(Sentinel2, _)

# Load Shapefile
shp = Shapefile.Table("data/landcover/landcover.shp") |> DataFrame

# Create Axes
fig = Figure();
ax1 = Axis(fig[1,1], xlabel="Wavelength (nm)", ylabel="Reflectance", title="Landsat Signatures");
ax2 = Axis(fig[2,1], xlabel="Wavelength (nm)", ylabel="Reflectance", title="Sentinel Signatures");

# Plot Signatures
plot_signatures!(ax1, Landsat8, landsat, shp, :MC_name; colors=cgrad(:tab10))
plot_signatures!(ax2, Sentinel2, sentinel, shp, :MC_name; colors=cgrad(:tab10))

# Add Legend
Legend(fig[:,2], ax1)

RemoteSensingToolbox.Spectral.plot_signatures — Method

plot_signatures(bandset::Type{<:AbstractBandset}, raster::AbstractRasterStack, shp, label::Symbol; colors=wong_colors())

Plot the spectral signatures for one or more land cover types.

Parameters

bandset: The sensor type to which the signatures belong.

raster: A RasterStack from which we want to extract the spectral signatures.

shp: A Tables.jl compatible object containing a :geometry column storing a GeoInterface.jl compatible geometry and a label column indicating the land cover type.

label: The column in shp corresponding to the land cover type.

colors: The color scheme used by the plot.

Example

# Read Landsat And Convert DNs To Reflectance
landsat = @pipe read_bands(Landsat8, "data/LC08_L2SP_043024_20200802_20200914_02_T1/") |> dn_to_reflectance(Landsat8, _)

# Load Shapefile
shp = Shapefile.Table("data/landcover/landcover.shp") |> DataFrame

# Plot Signatures
plot_signatures(Landsat8, landsat, shp, :MC_name)

Principal Component Analysis

Principal Component Analysis (PCA) is typically used to reduce the dimensionality of data. In the case of remote sensing, we are interested in reducing the number of bands we need to store while retaining as much information as possible. PCA rotates the bands into a new coordinate space where each band, called a principal component, is orthogonal to and uncorrelated with every other component. By convention, we order the bands in the transformed image in terms of their explained variance, such that the nth component accounts for more variance than any component after it.

Method	Description

fit_pca	Fit a PCA transformation to the provided data and return the analytical results.

forward_pca	Run a previously learned PCA transformation on a given Raster or RasterStack.

inverse_pca	Invert a PCA transformation on a previously transformed Raster or RasterStack.

projection	Return the projection matrix for a fitted PCA transformation.

explained_variance	Return the explained variance for each component of a fitted PCA transformation.

cumulative_variance	Return the cumulative variance for each component of a fitted PCA transformation.

RemoteSensingToolbox.Transformations.PCA — Type

Remotely sensed imagery typically consists of anywhere from a few to several hundred spectral bands. These bands are often highly correlated due to occupying similar spectral regions. Principal Component Analysis (PCA) is used in remote sensing to:

Create a smaller dataset from multiple bands, while retaining as much of the original spectral information as possible. The new image will consist of several uncorrelated PC bands.

Reveal complex relationships among spectral features.

Distinguish between characteristics that are prevalent in most bands and those that are specific to a few.

Example

julia> desis = read_bands(DESIS, "data/DESIS-HSI-L2A-DT0884573241_001-20200601T234520-V0210/");

julia> pca = fit_pca(desis, stats_fraction=0.1)
PCA(dimensions=235) 

Projection Matrix:
235×235 Matrix{Float32}:
  0.0001  0.0032   0.0016  -0.0094   0.0147  -0.0151  -0.0049   0.0163  …   0.0038  -0.0012   0.0008  -0.0032  -0.0007   0.0053
  0.0005  0.0099   0.0042  -0.0244   0.0517  -0.0335  -0.0185   0.0441     -0.0023  -0.0016  -0.0008   0.001    0.0003  -0.0004
  0.0003  0.015    0.0053  -0.037    0.133   -0.0443  -0.0271   0.1381     -0.0006   0.0004   0.0002  -0.0006   0.0003   0.0002
  0.0003  0.019    0.0071  -0.0385   0.1369  -0.0393  -0.0148   0.0949     -0.0008   0.0006  -0.0001  -0.0006   0.0001  -0.0006
  0.0003  0.0232   0.0073  -0.0419   0.1469  -0.037    0.0041   0.0839     -0.0013  -0.0025   0.0005   0.0007   0.0007  -0.0002
  0.0001  0.0267   0.0077  -0.0461   0.1713  -0.0325   0.0246   0.0861  …   0.0013  -0.0007  -0.0007   0.0024  -0.0012  -0.0028
  0.0001  0.0295   0.0083  -0.0476   0.1695  -0.0348   0.0319   0.0827     -0.0015  -0.0016  -0.0029   0.0004   0.0019  -0.0009
 -0.0001  0.0318   0.0086  -0.0482   0.17    -0.0352   0.0414   0.0784      0.0005  -0.002    0.0003   0.0021  -0.0003  -0.0022
  ⋮                                           ⋮                         ⋱            ⋮                                  
 -0.0663  0.0371  -0.1728   0.0196  -0.0508  -0.1394  -0.0054  -0.0226     -0.0003   0.0001   0.0007  -0.0009  -0.0002   0.0004
 -0.0658  0.0365  -0.1679   0.0204  -0.0717  -0.1474  -0.0087  -0.0193     -0.0006   0.0002   0.0004  -0.0      0.0004  -0.0001
 -0.0655  0.0352  -0.163    0.0193  -0.0767  -0.1511  -0.012   -0.0232     -0.0005   0.0002  -0.0004  -0.0001  -0.0002   0.0005
 -0.066   0.035   -0.1618   0.0193  -0.0859  -0.1503  -0.0055  -0.0177  …   0.0003   0.0005   0.001   -0.0      0.0003  -0.0002
 -0.067   0.035   -0.1619   0.019   -0.0745  -0.1466   0.0245  -0.0228     -0.0001   0.0002  -0.0002  -0.0001  -0.0002   0.0005
 -0.0679  0.0343  -0.1601   0.0176  -0.0721  -0.139    0.0328  -0.0286      0.0003   0.0003   0.0005  -0.0      0.0003   0.0002
 -0.0682  0.0337  -0.1588   0.0151  -0.0458  -0.1242   0.0549  -0.0315      0.0002  -0.0002  -0.0003   0.0002  -0.0003  -0.0005
 -0.0711  0.0343  -0.1601   0.012   -0.0612  -0.1804   0.2151  -0.0971      0.0004  -0.0001   0.0      0.0003   0.0     -0.0003

Importance of Components:
  Cumulative Variance: 0.8782  0.9493  0.9809  0.9869  0.9889  0.9906  0.9915  0.9922  0.9928  ...  1.0
  Explained Variance: 0.8782  0.0711  0.0316  0.006  0.0021  0.0017  0.0009  0.0007  0.0006  ...  0.0

julia> transformed = forward_pca(pca, desis, 12);

julia> size(transformed)
(1131, 1120, 12)

julia> recovered = inverse_pca(pca, transformed);

julia> size(recovered)
(1131, 1120, 235)

RemoteSensingToolbox.Transformations.fit_pca — Function

fit_pca(raster; method=:cov, stats_fraction=1.0)

Fit a Principal Component Analysis (PCA) transformation to the given AbstractRasterStack or AbstractRaster.

Parameters

raster: The AbstractRaster or AbstractRasterStack on which to fit the PCA transformation.

method: Either :cov or :cor, depending on whether we want to use the covariance or correlation matrix for computing the PCA rotation.

stats_fraction: The fraction of pixels to use when computing the covariance (or correlation) matrix. Values less than 1.0 will speed up computation at the cost of precision.

RemoteSensingToolbox.Transformations.forward_pca — Function

forward_pca(transformation::PCA, raster, components::Int)

Perform a forward Principal Component Analysis (PCA) transformation on the given raster, retaining only the specified number of components.

Parameters

transformation: A previously fitted PCA transformation.

raster: The AbstractRaster or AbstractRasterStack on which to perform the PCA transformation.

components: The number of bands to retain in the transformed image. All band numbers exceeding this value will be discarded.

RemoteSensingToolbox.Transformations.inverse_pca — Function

inverse_pca(transformation::PCA, raster::AbstractRaster)

Perform an inverse Principal Component Analysis (PCA) transformation to recover the original image.

Parameters

transformation: A previously fitted PCA transformation.

raster: An AbstractRaster representing a previously transformed image. The number of bands should be less than or equal to that of the original image.

RemoteSensingToolbox.Transformations.projection — Function

projection(x::PCA)

Return the projection matrix for the fitted PCA transformation.

projection(x::MNF)

Return the projection matrix for the fitted MNF transform.

RemoteSensingToolbox.Transformations.explained_variance — Function

explained_variance(x::PCA)

Return the explained variance associated with each principal component of the fitted PCA transform.

RemoteSensingToolbox.Transformations.cumulative_variance — Function

cumulative_variance(x::PCA)

Return the cumulative variance associated with each principal component of the fitted PCA transform.

Minimum Noise Fraction

The Minimum Noise Fraction (MNF) transformation is used to separate noise from data along the spectral dimension. This method is typically used with hyperspectral imagery, both as a means of dimension reduction and for noise removal. The transformed image will have its bands placed in descending order according to their Signal to Noise Ratio (SNR). The result is that the noise becomes concentrated in the higher bands, which can then be removed by either applying a standard image denoising algorithm or dropping them altogether.

Method	Description

fit_mnf	Fit an MNF transform to the provided data and return the analytical results.

forward_mnf	Run a previously learned MNF transformation on a given Raster or RasterStack.

inverse_mnf	Invert an MNF transformation on a previously transformed raster.

noise_cov	Return the noise covariance matrix for a fitted MNF transformation.

data_cov	Return the data covariance matrix for a fitted MNF transformation.

snr	Return the SNR for each principal component of a fitted MNF transformation.

eigenvalues	Return eigenvalues for each principal component of a fitted MNF transformation.

cumulative_snr	Return the cumulative SNR for each PC of a fitted MNF transformation.

cumulative_eigenvalues	Return the cumulative eigenvalues for each PC of a fitted MNF transformation.

RemoteSensingToolbox.Transformations.MNF — Type

The Minimum Noise Fraction (MNF) transform is a linear transformation used to reduce the spectral dimensionality of image data and segregate noise. MNF consists of two separate principal component rotations. The first rotation uses the principal components of the noise covariance matrix to decorrelate and rescale the noise (a process known as noise whitening), resulting in a transformed image in which the noise has unit variance and no band-to-band correlations. The second rotation is a standard PCA rotation applied to the noise-whitened image.

The bands in the transformed image will be ordered according to their Signal to Noise Ratio (SNR), with the highest SNR being placed first. The result is that the noise becomes concentrated in the higher bands. Thus, the transform can be used to separate noise from data by performing a forward transform, determining which bands contain coherent images, and running an inverse transform after either discarding or denoising the remaining bands. The number of bands to keep in the inverse transform can be determined by a number of methods. The simplest approach is to look at the sorted transformed bands and threshold at the band where no recognizable features can be observed. An alternative method is to threshold at a desired cumulative SNR.

Example

julia> desis = read_bands(DESIS, "data/DESIS-HSI-L2A-DT0884573241_001-20200601T234520-V0210/");

julia> roi = @view desis[X(1019:1040), Y(550:590)];

julia> mnf = fit_mnf(desis, roi)
MNF(dimensions=235) 

Projection Matrix:
235×235 Matrix{Float32}:
 -2.135    2.3502   0.3612   0.5912   0.5217  -0.0917   0.0043  …   0.0002  -0.0001  -0.0004  -0.0004  -0.0      0.0004
 -0.0959   0.0422  -0.0047  -0.2362  -0.3962  -0.2313  -0.1685      0.0001   0.0004   0.0002   0.0003   0.0001  -0.0001
  0.0043   0.0058  -0.0032   0.0023  -0.0061  -0.0048   0.0028     -0.0001   0.0001  -0.0001  -0.0      0.0      0.0001
  0.0039   0.002   -0.002    0.0012  -0.0032  -0.004    0.006      -0.0006   0.0002  -0.0001  -0.0     -0.0003  -0.0
  0.0024   0.0018  -0.003   -0.0008   0.0038  -0.0003   0.0057      0.0009  -0.0007   0.0002  -0.0012   0.0012  -0.0
  0.0019  -0.003    0.0038  -0.002   -0.0001  -0.0003  -0.0     …   0.0021   0.0009  -0.0004   0.0014  -0.0011  -0.0006
  0.0047   0.0055   0.006   -0.0014  -0.0011   0.0021   0.0053     -0.0035   0.0006   0.0002  -0.0026   0.0014   0.0019
  0.0072   0.0042   0.0012   0.0016   0.0011  -0.002   -0.001       0.0     -0.0026   0.0012   0.0034  -0.0006  -0.0009
  ⋮                                            ⋮                ⋱            ⋮                                  
 -0.0004  -0.0012   0.0007   0.0006  -0.0      0.0002   0.0005      0.0005   0.0002  -0.0004  -0.0002  -0.0001   0.0001
 -0.0014  -0.0005  -0.0005   0.0019  -0.0002  -0.0005   0.0017      0.0003   0.0005   0.0      0.0004   0.0005   0.0
 -0.0004   0.0008  -0.0003   0.0013   0.0004   0.0014   0.0004      0.0002  -0.0001   0.0002   0.0001  -0.0002  -0.0002
 -0.0008  -0.0015   0.0021  -0.0004  -0.0004   0.0012   0.0006  …  -0.0011   0.0002  -0.0007   0.0001   0.0005  -0.0001
  0.0008  -0.0004   0.0009   0.0019  -0.0022  -0.0014   0.0013      0.0003  -0.0001  -0.0003  -0.0      0.0003   0.0002
  0.0005   0.0006  -0.0022  -0.0003   0.0     -0.0022   0.0016     -0.001    0.0002  -0.0003   0.0003  -0.0005   0.0005
  0.001   -0.0005  -0.0007   0.0025  -0.0019   0.0005   0.0015      0.0002  -0.0005   0.0     -0.0005   0.0002  -0.0002
 -0.0003   0.0002  -0.0021  -0.0008  -0.0012   0.0003   0.0003     -0.0001  -0.0      0.0001   0.0      0.0      0.0

Component Statistics:
  Eigenvalues: 7975.439  4040.6348  2092.866  717.8178  468.5496  247.5029  202.2003  176.8452  87.3302  ...  0.3602
  Cumulative Eigenvalues: 0.4779  0.72  0.8454  0.8884  0.9165  0.9313  0.9434  0.954  0.9593  ...  1.0
  Explained SNR: 7974.4385  4039.6353  2091.8635  716.8176  467.55  246.5022  201.1985  175.845  86.3301  ...  -0.6399
  Cumulative SNR: 0.4839  0.729  0.856  0.8995  0.9278  0.9428  0.955  0.9657  0.9709  ...  0.9985

julia> transformed = forward_mnf(mnf, desis, 12);

julia> size(transformed)
(1131, 1120, 12)

julia> recovered = inverse_mnf(mnf, transformed);

julia> size(recovered)
(1131, 1120, 235)

RemoteSensingToolbox.Transformations.fit_mnf — Function

fit_mnf(raster, noise_sample; smooth=true)

Fit a Minimum Noise Fraction (MNF) transformation to the given AbstractRasterStack or AbstractRaster.

Parameters

raster: The AbstractRaster or AbstractRasterStack on which to fit the MNF transformation.

noise_sample: A homogenous (spectrally uniform) region extracted from raster for calculating the noise covariance matrix.

smooth: The MNF transform cannot be computed if any band in noise_sample has zero variance. To correct this, you may wish to introduce a small smoothing term (true by default).

RemoteSensingToolbox.Transformations.forward_mnf — Function

forward_mnf(transformation::MNF, raster, components::Int)
forward_mnf(transformation::MNF, sigs::Matrix, components::Int)

Perform a forward Minimum Noise Fraction (MNF) rotation on the given raster or signatures, retaining only the specified number of components.

Parameters

transformation: A previously fitted MNF transformation.

raster: The AbstractRaster or AbstractRasterStack on which to perform the MNF transformation.

sigs: An n x b matrix of spectral signatures where n is the number of signatures and b is the number of bands.

components: The number of bands to retain in the transformed image. All band numbers exceeding this value will be discarded.

RemoteSensingToolbox.Transformations.inverse_mnf — Function

inverse_mnf(transformation::MNF, raster::AbstractRaster)
inverse_mnf(transformation::MNF, sigs::Matrix)

Perform an inverse Minimum Noise Fraction (MNF) transformation to recover the original image or signatures.

Parameters

transformation: A previously fitted MNF transformation.

raster: An AbstractRaster representing a previously transformed image. The number of bands should be less than or equal to that of the original image.

sigs: An n x p matrix of transformed signatures where n is the number of signatures and p is the number of retained components.

RemoteSensingToolbox.Transformations.noise_cov — Function

noise_cov(x::MNF)

Return the noise covariance matrix for the fitted MNF transform.

RemoteSensingToolbox.Transformations.data_cov — Function

data_cov(x::MNF)

Return the data covariance matrix for the fitted MNF transform.

RemoteSensingToolbox.Transformations.snr — Function

snr(x::MNF)

Return the estimated SNR associated with each principal components of the fitted MNF transform.

RemoteSensingToolbox.Transformations.eigenvalues — Function

eigenvalues(x::MNF)

Return the eigenvalues associated with each principal components of the fitted MNF transform.

RemoteSensingToolbox.Transformations.cumulative_snr — Function

cumulative_snr(x::MNF)

Return the cumulative SNR associated with each principal components of the fitted MNF transform.

RemoteSensingToolbox.Transformations.cumulative_eigenvalues — Function

cumulative_eigenvalues(x::MNF)

Return the cumulative eigenvalues associated with each principal components of the fitted MNF transform.

Utilities

RemoteSensingToolbox provides several utility functions for modifying and processing remotely sensed data.

Method	Description

mask_pixels	Remove all pixels in a raster covered by a provided mask.

mask_pixels!	The mutating form of mask_pixels.

tocube	Convert a multi-layer RasterStack into a multiband Raster.

RemoteSensingToolbox.create_tiles — Method

create_tiles(raster, tile::Tuple{Int,Int}; stride=tile)

Slice the given raster into tiles with size tile.

Parameters

raster: The raster to be cut into tiles.

tile: The size of the generated tiles in terms of width x height.

stride: The distance between the top-left corner of each tile. Is equal to tile by default, which produces non-overlapping tiles.

RemoteSensingToolbox.encode — Method

encode(raster::AbstractRaster, type; missingval=typemax(type))

Encodes a raster into the provided type.

Parameters

raster: The AbstractRaster or AbstractRasterStack to be encoded.

type: The type to encode as. Common values include UInt16, Int16, and Float32.

missingval: The value to use to denote missing pixels. Set to the maximum value of type by default.

RemoteSensingToolbox.mask_pixels! — Method

mask_pixels!(raster, mask; invert_mask=false)

Drop pixels from a raster according to a given mask. The mask and raster must have the same extent and size.

Parameters

raster: The raster to be masked.

mask: A mask defining which pixels we want to drop. By default, we drop pixels corresponding to mask values of 1.

invert_mask: Treat mask values of 1 as 0 and vice-versa.

RemoteSensingToolbox.mask_pixels — Method

mask_pixels(raster, mask; invert_mask=false)

Drop pixels from a raster according to a given mask. The mask and raster must have the same extent and size.

Parameters

raster: The raster to be masked.

mask: A mask defining which pixels we want to drop. By default, we drop pixels corresponding to mask values of 1.

invert_mask: Treat mask values of 1 as 0 and vice-versa.

RemoteSensingToolbox.tocube — Method

tocube(rs::RasterStack; layers=names(rs))
tocube(rs::AbstractBandset; layers=names(rs))

Transform the multi-layer RasterStack to a multi-band raster.

Parameters

X: The RasterStack or AbstractBandset to be transformed into a multi-band raster.

layers: The layers to include in the new raster.

Example

julia> landsat = read_bands(Landsat8, "LC08_L2SP_043024_20200802_20200914_02_T1/");

julia> tocube(landsat)
7821×7921×7 Raster{Float32,3} B1 with dimensions: 
  X Projected{Float64} LinRange{Float64}(493785.0, 728385.0, 7821) ForwardOrdered Regular Points crs: WellKnownText,
  Y Projected{Float64} LinRange{Float64}(5.84638e6, 5.60878e6, 7921) ReverseOrdered Regular Points crs: WellKnownText,
  Band Categorical{Int64} 1:7 ForwardOrdered
extent: Extent(X = (493785.0, 728385.0), Y = (5.608785e6, 5.846385e6), Band = (1, 7))
missingval: 0.0f0
crs: PROJCS["WGS 84 / UTM zone 11N",GEOGCS["WGS 84",DATUM["WGS_1984",SPHEROID["WGS 84",6378137,298.257223563,AUTHORITY["EPSG","7030"]],AUTHORITY["EPSG","6326"]],PRIMEM["Greenwich",0,AUTHORITY["EPSG","8901"]],UNIT["degree",0.0174532925199433,AUTHORITY["EPSG","9122"]],AUTHORITY["EPSG","4326"]],PROJECTION["Transverse_Mercator"],PARAMETER["latitude_of_origin",0],PARAMETER["central_meridian",-117],PARAMETER["scale_factor",0.9996],PARAMETER["false_easting",500000],PARAMETER["false_northing",0],UNIT["metre",1,AUTHORITY["EPSG","9001"]],AXIS["Easting",EAST],AXIS["Northing",NORTH],AUTHORITY["EPSG","32611"]]
parent:
[:, :, 1]
           5.84638e6  5.84636e6  5.84632e6  5.8463e6  5.84626e6  …  5.60894e6  5.6089e6  5.60888e6  5.60884e6  5.60882e6  5.60878e6
 493785.0  0.0        0.0        0.0        0.0       0.0           0.0        0.0       0.0        0.0        0.0        0.0
      ⋮                                               ⋮          ⋱                                             ⋮          
 728355.0  0.0        0.0        0.0        0.0       0.0           0.0        0.0       0.0        0.0        0.0        0.0
 728385.0  0.0        0.0        0.0        0.0       0.0           0.0        0.0       0.0        0.0        0.0        0.0
[and 6 more slices...]