RemoteSensingToolbox

RemoteSensingToolbox is a pure Julia package intended to provide a collection of tools for visualizing, manipulating, and interpreting remotely sensed imagery.

Bandsets

Bandsets are julia types that encode the sensor-specific information needed for many methods in RemoteSensingToolbox to work without the need for tedious details provided by the end user. Bandsets are provided for several common sensors including Sentinel-2, Landsat 7, and Landsat 8. Unsupported sensors can be added by defining a custom type, which should be a sub-type of AbstractBandset, and implementing the following interface:

Method	Description	Required

bands	Return the band names in order from shortest to longest wavelength.	yes

wavelengths	Return the central wavelengths for all bands from shortest to longest.	yes

blue	Return the blue band for the given sensor.	yes

green	Return the green band for the given sensor.	yes

red	Return the red band for the given sensor.	yes

nir	Return the nir band for the given sensor.	yes

swir1	Return the swir1 band for the given sensor.	yes

swir2	Return the swir2 band for the given sensor.	yes

parse_band	Parses the band from a given filename. Used by read_bands.	no

read_qa	Reads the QA or scene classification file from the provided file or directory.	no

dn_to_reflectance	Decodes digital numbers to reflectance.	no

RemoteSensingToolbox.Bandsets.AbstractBandset — Type

The supertype of all sensor types. Provides sensor-specific information to many RemoteSensingToolbox methods.

Example Minimal Implementation

struct Landsat8 <: AbstractBandset end

bands(::Type{Landsat8}) = [:B1, :B2, :B3, :B4, :B5, :B6, :B7]

wavelengths(::Type{Landsat8}) = [443, 483, 560, 660, 865, 1650, 2220]

blue(::Type{Landsat8}) = :B2

green(::Type{Landsat8}) = :B3

red(::Type{Landsat8}) = :B4

nir(::Type{Landsat8}) = :B5

swir1(::Type{Landsat8}) = :B6

swir2(::Type{Landsat8}) = :B7

RemoteSensingToolbox.Bandsets.DESIS — Type

Implements the AbstractBandset interface for Landsat 7.

RemoteSensingToolbox.Bandsets.Landsat8 — Type

Implements the AbstractBandset interface for Landsat 8.

RemoteSensingToolbox.Bandsets.Landsat7 — Type

Implements the AbstractBandset interface for Landsat 7.

RemoteSensingToolbox.Bandsets.Sentinel2 — Type

Implements the AbstractBandset interface for Sentinel 2.

RemoteSensingToolbox.Bandsets.red — Function

red(::Type{AbstractBandset})
red(raster, ::Type{AbstractBandset})

Return the red band for the given sensor.

RemoteSensingToolbox.Bandsets.green — Function

green(::Type{AbstractBandset})
green(raster, ::Type{AbstractBandset})

Return the green band for the given sensor.

RemoteSensingToolbox.Bandsets.blue — Function

blue(::Type{AbstractBandset})
blue(raster, ::Type{AbstractBandset})

Return the blue band for the given sensor.

RemoteSensingToolbox.Bandsets.nir — Function

nir(::Type{AbstractBandset})
nir(raster, ::Type{AbstractBandset})

Return the nir band for the given sensor.

RemoteSensingToolbox.Bandsets.swir1 — Function

swir1(::Type{AbstractBandset})
swir1(raster, ::Type{AbstractBandset})

Return the swir1 band for the given sensor.

RemoteSensingToolbox.Bandsets.swir2 — Function

swir2(::Type{AbstractBandset})
swir2(raster, ::Type{AbstractBandset})

Return the swir2 band for the given sensor.

RemoteSensingToolbox.Bandsets.bands — Function

bands(::Type{AbstractBandset})

Return the band names in order from shortest to longest wavelength.

RemoteSensingToolbox.Bandsets.wavelengths — Function

wavelengths(::Type{AbstractBandset})

Return the central wavelengths for all bands in order from shortest to longest.

RemoteSensingToolbox.Bandsets.wavelength — Function

wavelength(::Type{AbstractBandset}, band::Symbol)

Return the central wavelength for the corresponding band.

RemoteSensingToolbox.Bandsets.parse_band — Function

parse_band(::Type{AbstractBandset}, filename::String)

Parses the band name from the given file path.

Returns either the band as a Symbol or nothing if no band could be parsed.

If the file is a multi-band raster, returns the names of all bands as a Vector{Symbol}.

RemoteSensingToolbox.Bandsets.read_bands — Function

function read_bands(bandset::Type{AbstractBandset}, dir::String)

Read the bands from the given directory into a RasterStack.

Parameters

bandset: The sensor type to which the bands belong.

dir: The directory in which the bands can be read.

Example

julia> landsat = read_bands(Landsat8, "data/LC08_L2SP_043024_20200802_20200914_02_T1/")
RasterStack with dimensions: 
  X Projected{Float64} LinRange{Float64}(493785.0, 728385.0, 7821) ForwardOrdered Regular Points crs: WellKnownText,
  Y Projected{Float64} LinRange{Float64}(5.84638e6, 5.60878e6, 7921) ReverseOrdered Regular Points crs: WellKnownText
and 7 layers:
  :B1 UInt16 dims: X, Y (7821×7921)
  :B2 UInt16 dims: X, Y (7821×7921)
  :B3 UInt16 dims: X, Y (7821×7921)
  :B4 UInt16 dims: X, Y (7821×7921)
  :B5 UInt16 dims: X, Y (7821×7921)
  :B6 UInt16 dims: X, Y (7821×7921)
  :B7 UInt16 dims: X, Y (7821×7921)

RemoteSensingToolbox.Bandsets.read_qa — Function

read_qa(bandset::Type{AbstractBandSet}, src::String)

Read and decode the quality assurance mask for the given AbstractBandset.

Parameters

bandset: A subtype of AbstractBandset.

src: Either a directory containing the quality assurance mask named according to standard conventions or the file itself.

Returns

The decoded quality assurance mask as a RasterStack. Masked values are encoded as 1, non-masked values as 0, and missing values as 255.

Example

julia> qa = read_qa(Landsat8, "data/LC08_L2SP_043024_20200802_20200914_02_T1/")
RasterStack with dimensions: 
  X Projected{Float64} LinRange{Float64}(493785.0, 728385.0, 7821) ForwardOrdered Regular Points crs: WellKnownText,
  Y Projected{Float64} LinRange{Float64}(5.84638e6, 5.60878e6, 7921) ReverseOrdered Regular Points crs: WellKnownText
and 7 layers:
  :dilated_cloud UInt8 dims: X, Y (7821×7921)
  :cirrus        UInt8 dims: X, Y (7821×7921)
  :cloud         UInt8 dims: X, Y (7821×7921)
  :cloud_shadow  UInt8 dims: X, Y (7821×7921)
  :snow          UInt8 dims: X, Y (7821×7921)
  :clear         UInt8 dims: X, Y (7821×7921)
  :water         UInt8 dims: X, Y (7821×7921)

RemoteSensingToolbox.Bandsets.dn_to_reflectance — Function

dn_to_reflectance(bandset::Type{AbstractBandset}, raster; clamp_values=false)

Transform the raster from Digital Numbers (DN) to reflectance.

Parameters

bandset: A subtype of AbstractBandset.

raster: The AbstractRasterStack or AbstractRaster to be converted to reflectance.

clamp_values: Indicates whether to clamp reflectances into the range (0.0, 1.0] (default = false).

Example

landsat = read(Landsat8, "LC08_L2SP_043024_20200802_20200914_02_T1/")
landsat_sr = dn_to_reflectance(Landsat8, landsat)

Visualization

RemoteSensingToolbox.Agriculture — Type

Agriculture band composite.

RemoteSensingToolbox.ColorInfrared — Type

Color infrared band composite.

RemoteSensingToolbox.Geology — Type

Geology band composite.

RemoteSensingToolbox.SWIR — Type

SWIR band composite.

RemoteSensingToolbox.TrueColor — Type

True color band composite.

RemoteSensingToolbox.visualize — Method

visualize(g::AbstractRaster; lower=0.02, upper=0.98)
visualize(r::AbstractRaster, g::AbstractRaster, b::AbstractRaster; lower=0.02, upper=0.98)
visualize(img::AbstractBandSet, ::Type{TrueColor{AbstractBandset}}; kwargs...)
visualize(img::AbstractBandSet, ::Type{ColorInfrared{AbstractBandset}}; kwargs...)
visualize(img::AbstractBandSet, ::Type{SWIR{AbstractBandset}}; kwargs...)
visualize(img::AbstractBandSet, ::Type{Agriculture{AbstractBandset}}; kwargs...)
visualize(img::AbstractBandSet, ::Type{Geology{AbstractBandset}}; kwargs...)

Visualize a satellite image after applying a histogram stretch. Returns either an RGB or grayscale image compatible with the Images.jl ecosystem.

A number of band combinations are supported for types implementing the AbstractBandSet interface.

Example 1

landsat = read(Landsat8, "LC08_L2SP_043024_20200802_20200914_02_T1/")
img = mndwi(landsat, Landsat8) |> visualize
save("mndwi.png", img)

Example 2

landsat = read(Landsat8, "LC08_L2SP_043024_20200802_20200914_02_T1/")
img = visualize(landsat, TrueColor{Landsat8}; upper=0.90)
save("truecolor.png", img)

Preprocessing

RemoteSensingToolbox.create_tiles — Method

create_tiles(raster, tile::Tuple{Int,Int}; stride=tile)

Slice the given raster into tiles with size tile.

Parameters

raster: The raster to be cut into tiles.

tile: The size of the generated tiles in terms of width x height.

stride: The distance between the top-left corner of each tile. Is equal to tile by default, which produces non-overlapping tiles.

RemoteSensingToolbox.mask_pixels! — Method

mask_pixels!(raster, mask; invert_mask=false)

Drop pixels from a raster according to a given mask. The mask and raster must have the same extent and size.

Parameters

raster: The raster to be masked.

mask: A mask defining which pixels we want to drop. By default, we drop pixels corresponding to mask values of 1.

invert_mask: Treat mask values of 1 as 0 and vice-versa.

RemoteSensingToolbox.mask_pixels — Method

mask_pixels(raster, mask; invert_mask=false)

Drop pixels from a raster according to a given mask. The mask and raster must have the same extent and size.

Parameters

raster: The raster to be masked.

mask: A mask defining which pixels we want to drop. By default, we drop pixels corresponding to mask values of 1.

invert_mask: Treat mask values of 1 as 0 and vice-versa.

RemoteSensingToolbox.tocube — Method

tocube(rs::RasterStack; layers=names(rs))
tocube(rs::AbstractBandset; layers=names(rs))

Transform the multi-layer RasterStack to a multi-band raster.

Parameters

X: The RasterStack or AbstractBandset to be transformed into a multi-band raster.

layers: The layers to include in the new raster.

Example

julia> landsat = Landsat8("LC08_L2SP_043024_20200802_20200914_02_T1");
julia> tocube(landsat)
7821×7921×7 Raster{Float32,3} B1 with dimensions: 
  X Projected{Float64} LinRange{Float64}(493785.0, 728385.0, 7821) ForwardOrdered Regular Points crs: WellKnownText,
  Y Projected{Float64} LinRange{Float64}(5.84638e6, 5.60878e6, 7921) ReverseOrdered Regular Points crs: WellKnownText,
  Band Categorical{Int64} 1:7 ForwardOrdered
extent: Extent(X = (493785.0, 728385.0), Y = (5.608785e6, 5.846385e6), Band = (1, 7))
missingval: 0.0f0
crs: PROJCS["WGS 84 / UTM zone 11N",GEOGCS["WGS 84",DATUM["WGS_1984",SPHEROID["WGS 84",6378137,298.257223563,AUTHORITY["EPSG","7030"]],AUTHORITY["EPSG","6326"]],PRIMEM["Greenwich",0,AUTHORITY["EPSG","8901"]],UNIT["degree",0.0174532925199433,AUTHORITY["EPSG","9122"]],AUTHORITY["EPSG","4326"]],PROJECTION["Transverse_Mercator"],PARAMETER["latitude_of_origin",0],PARAMETER["central_meridian",-117],PARAMETER["scale_factor",0.9996],PARAMETER["false_easting",500000],PARAMETER["false_northing",0],UNIT["metre",1,AUTHORITY["EPSG","9001"]],AXIS["Easting",EAST],AXIS["Northing",NORTH],AUTHORITY["EPSG","32611"]]
parent:
[:, :, 1]
           5.84638e6  5.84636e6  5.84632e6  5.8463e6  5.84626e6  …  5.60894e6  5.6089e6  5.60888e6  5.60884e6  5.60882e6  5.60878e6
 493785.0  0.0        0.0        0.0        0.0       0.0           0.0        0.0       0.0        0.0        0.0        0.0
      ⋮                                               ⋮          ⋱                                             ⋮          
 728355.0  0.0        0.0        0.0        0.0       0.0           0.0        0.0       0.0        0.0        0.0        0.0
 728385.0  0.0        0.0        0.0        0.0       0.0           0.0        0.0       0.0        0.0        0.0        0.0
[and 6 more slices...]

Land Cover Indices

RemoteSensingToolbox.mndwi — Method

mndwi(green::AbstractRaster, swir::AbstractRaster)
mndwi(stack::AbstractRasterStack, ::Type{AbstractBandset})

Compute the Modified Normalised Difference Water Index (Xu 2006).

MNDWI = (green - swir) / (green + swir)

RemoteSensingToolbox.nbri — Method

nbri(nir::AbstractRaster, swir2::AbstractRaster)
nbri(stack::AbstractRasterStack, ::Type{AbstractBandset})

Compute the Normalized Burn Ratio Index.

NBRI is used to emphasize burned areas.

NBRI = (nir - swir2) / (nir + swir2)

RemoteSensingToolbox.ndbi — Method

ndbi(swir1::AbstractRaster, nir::AbstractRaster)
ndbi(stack::AbstractRasterStack, ::Type{AbstractBandset})

Compute the The Normalized Difference Built-up Index

NDBI is used to emphasize urban and built-up areas.

NDBI = (swir1 - nir) / (swir1 + nir)

RemoteSensingToolbox.ndmi — Method

ndmi(nir::AbstractRaster, swir1::AbstractRaster)
ndmi(stack::AbstractRasterStack, ::Type{AbstractBandset})

Compute the Normalized Difference Moisture Index.

NDMI is sensitive to the moisture levels in vegetation. It is used to monitor droughts and fuel levels in fire-prone areas.

NDMI = (nir - swir1) / (nir + swir1)

RemoteSensingToolbox.ndvi — Method

ndvi(nir::AbstractRaster, red::AbstractRaster)
ndvi(stack::AbstractRasterStack, ::Type{AbstractBandset})

Compute the Normalized Difference Vegetation Index.

NDVI = (nir - red) / (nir + red)

RemoteSensingToolbox.ndwi — Method

ndwi(green::AbstractRaster, nir::AbstractRaster)
ndwi(stack::AbstractRasterStack, ::Type{AbstractBandset})

Compute the Normalized Difference Water Index (McFeeters 1996).

NDWI = (green - nir) / (green + nir)

RemoteSensingToolbox.savi — Method

savi(nir::AbstractRaster, red::AbstractRaster; L=0.33)
savi(stack::AbstractRasterStack, ::Type{AbstractBandset}; L=0.33)

Compute the Soil Adjusted Vegetation Index (Huete 1988).

SAVI is a vegetation index which attempts to minimize soil brightness influences by introducing a soil-brightness correction factor (L).

L represents the amount of green vegetation cover, which is set to 0.33 by default.

SAVI = ((nir - red) / (nir + red + L)) * (1 + L)

Spectral Analysis

RemoteSensingToolbox.Spectral.extract_signatures — Method

extract_signatures(stack::AbstractRasterStack, shp, label::Symbol; label_name=nothing)

Extract signatures from the given RasterStack or AbstractSensor within regions specified by a given shapefile.

Parameters

stack: The RasterStack from which to extract spectral signatures.

shp: A Tables.jl compatible object containing a :geometry column storing a GeoInterface.jl compatible geometry and a label column indicating the land cover type.

label: The column in shp in which the land cover class is stored.

Returns

A DataFrame consisting of rows for each extracted signature and columns storing the respective bands and land cover type.

Example

julia> landsat = @pipe read_bands(Landsat8, "data/LC08_L2SP_043024_20200802_20200914_02_T1/") |> dn_to_reflectance(Landsat8, _)

julia> shp = Shapefile.Table("data/landcover/landcover.shp") |> DataFrame

julia> extract_signatures(landsat, shp, :C_name)
1925×8 DataFrame
  Row │ B1         B2         B3        B4        B5        B6        B7        label     
      │ Float32    Float32    Float32   Float32   Float32   Float32   Float32   String    
──────┼───────────────────────────────────────────────────────────────────────────────────
    1 │ 0.057235   0.10547    0.188932  0.24847   0.513405  0.315323  0.166107  Hail Scar
    2 │ 0.0574     0.105415   0.188712  0.24946   0.513735  0.317908  0.167593  Hail Scar
    3 │ 0.0584175  0.107175   0.19182   0.253887  0.515577  0.319585  0.168087  Hail Scar
    4 │ 0.0583625  0.10723    0.190995  0.253393  0.514615  0.317715  0.167097  Hail Scar
    5 │ 0.05806    0.10745    0.189895  0.250368  0.50774   0.315048  0.165475  Hail Scar
  ⋮   │     ⋮          ⋮         ⋮         ⋮         ⋮         ⋮         ⋮          ⋮
 1921 │ 0.077145   0.112097   0.17081   0.19677   0.25705   0.320877  0.28015   Built Up
 1922 │ 0.10712    0.136765   0.19666   0.225068  0.268655  0.33284   0.302288  Built Up
 1923 │ 0.0856425  0.123235   0.185853  0.211675  0.273     0.373623  0.351072  Built Up
 1924 │ 0.088145   0.119578   0.17059   0.204442  0.27487   0.355857  0.325553  Built Up
 1925 │ 0.0518725  0.0816275  0.148673  0.15189   0.28774   0.309218  0.29786   Built Up
                                                                         1915 rows omitted

RemoteSensingToolbox.Spectral.plot_signatures! — Method

plot_signatures!(ax, bandset::Type{<:AbstractBandset}, sigs::DataFrame; label=:label, colors=wong_colors())

Plot spectral signatures for each land cover type specified in a given shapefile by mutating a Makie.Axis object.

Accepts the same keywords as Makie.lines!.

Parameters

ax: The Makie.Axis into which we want to draw our plot.

bandset: The sensor type to which the signatures belong.

sigs: A DataFrame of signatures with columns corresponding to bands and labels.

label: The column of sigs in which the land cover type is stored (default = :label).

colors: The color scheme used by the plot.

Example

# Read Images And Convert DNs To Reflectance
landsat = @pipe read_bands(Landsat8, "data/LC08_L2SP_043024_20200802_20200914_02_T1/") |> dn_to_reflectance(Landsat8, _)
sentinel = @pipe read_bands(Sentinel2, "data/T11UPT_20200804T183919/R60m/") |> dn_to_reflectance(Sentinel2, _)

# Load Shapefile
shp = Shapefile.Table("data/landcover/landcover.shp") |> DataFrame

# Extract Signatures
landsat_sigs = extract_signatures(landsat, shp, :MC_name) |> summarize_signatures
sentinel_sigs = extract_signatures(sentinel, shp, :MC_name) |> summarize_signatures

# Create Axes
fig = Figure();
ax1 = Axis(fig[1,1], xlabel="Wavelength (nm)", ylabel="Reflectance", title="Landsat Signatures");
ax2 = Axis(fig[2,1], xlabel="Wavelength (nm)", ylabel="Reflectance", title="Sentinel Signatures");

# Plot Signatures
plot_signatures!(ax1, Landsat8, landsat_sigs; colors=cgrad(:tab10))
plot_signatures!(ax2, Sentinel2, sentinel_sigs; colors=cgrad(:tab10))

# Add Legend
Legend(fig[:,2], ax1)

RemoteSensingToolbox.Spectral.plot_signatures — Method

plot_signatures(bandset::Type{<:AbstractBandset}, sigs::DataFrame; label=:label, colors=wong_colors())

Plot the spectral signatures for one or more land cover types.

Parameters

bandset: The sensor type to which the signatures belong.

sigs: A DataFrame of signatures with columns corresponding to bands and labels.

label: The column of sigs in which the land cover type is stored (default = :label).

colors: The color scheme used by the plot.

Example

# Read Landsat And Convert DNs To Reflectance
landsat = @pipe read_bands(Landsat8, "data/LC08_L2SP_043024_20200802_20200914_02_T1/") |> dn_to_reflectance(Landsat8, _)

# Load Shapefile
shp = Shapefile.Table("data/landcover/landcover.shp") |> DataFrame

# Extract Signatures
sigs = extract_signatures(landsat, shp, :C_name) |> summarize_signatures

# Plot Signatures
plot_signatures(Landsat8, sigs)

RemoteSensingToolbox.Spectral.summarize_signatures — Function

summarize_signatures([reducer], sigs::DataFrame, [label])

Summarize a collection of spectral signatures by grouping according to their land cover type, then reducing each group to a single summary statistic.

Parameters

reducer: A function that reduces a collection of values to a single statistic (defaults to mean).

sigs: A DataFrame of signatures, where each row contains the band measurements for a single pixel and the corresponding land cover type.

label: The column in sigs which stores the land cover type (default is :label).

Returns

A DataFrame consisting of rows for each land cover type's summarized signature.

Example

julia> landsat = @pipe read_bands(Landsat8, "data/LC08_L2SP_043024_20200802_20200914_02_T1/") |> dn_to_reflectance(Landsat8, _)

julia> shp = Shapefile.Table("data/landcover/landcover.shp") |> DataFrame

julia> sigs = extract_signatures(landsat, shp, :C_name)

julia> summarize_signatures(sigs, :label)
7×8 DataFrame
 Row │ label       B1           B2          B3         B4          B5        B6          B7         
     │ String      Float32      Float32     Float32    Float32     Float32   Float32     Float32    
─────┼──────────────────────────────────────────────────────────────────────────────────────────────
   1 │ Hail Scar   0.0617346    0.107954    0.188092   0.247114    0.508847  0.322815    0.173574
   2 │ Bare Earth  0.0483927    0.0539072   0.0773476  0.0883738   0.231219  0.307681    0.199805
   3 │ Road        0.0396956    0.0530674   0.0933762  0.0952315   0.245672  0.205506    0.142639
   4 │ Lake        0.000517442  0.00360272  0.0132789  0.00628491  0.031176  0.00697667  0.00377249
   5 │ Trees       0.0182846    0.0204864   0.0404257  0.0245542   0.319328  0.139772    0.0585171
   6 │ Vegetation  0.00442494   0.015797    0.0789011  0.0464686   0.49601   0.0964562   0.0407997
   7 │ Built Up    0.0892711    0.118764    0.177746   0.200785    0.293111  0.33917     0.304133

Transformations

RemoteSensingToolbox.Transformations.AbstractTransformation — Type

The supertype of all transformations. Subtypes are expected to implement the fit and transform methods.

RemoteSensingToolbox.Transformations.Normalize — Type

A struct for storing the parameters necessary to perform a normalization transformation.

RemoteSensingToolbox.Transformations.PCA — Type

A struct for storing the parameters necessary to perform a PCA transformation.

RemoteSensingToolbox.Transformations.fit_transform — Method

fit_transform(transformation::Type{AbstractTransformation}, raster; kwargs...)
fit_transform(transformation::Type{Normalize}, raster)
fit_transform(transformation::Type{PCA}, raster; components=nbands(raster), method=:cov, stats_fraction=1.0)

Fit the specified transformation to the given AbstractRasterStack or AbstractRaster.

Parameters

transformation: The transformation we want to fit.

raster: The AbstractRaster or AbstractRasterStack on which to fit the specified transformation.

Supported Transformations

Normalize: Scales and shifts the data so that each band has a mean of 0 and a standard deviation of 1.

PCA: Rotates each pixel into a new orthogonal color-space, which may have fewer dimensions than the original.

PCA Parameters

components: The number of principal components to use.

method: One of either :cov or :cor, depending on whether we want to run PCA on the covariance or the correlation matrix.

stats_fraction: The fraction of pixels to use in the calculation. Setting stats_fraction < 1 will produce faster but less accurate results.

RemoteSensingToolbox.Transformations.inverse_transform — Method

inverse_transform(transformation::AbstractTransformation, raster)
inverse_transform(transformation::Normalize, raster; output_type=Float32)
inverse_transform(transformation::PCA, raster; output_type=Float32)

Undo a previously applied transformation.

Parameters

transformation: Some previously applied transformation that we want to reverse.

raster: A previously transformed AbstractRaster or AbstractRasterStack from which we want to recover the original.

Supported Transformations

Normalize: Scales and shifts the data so that each band has a mean of 0 and a standard deviation of 1.

PCA: Rotates each pixel into a new orthogonal color-space, which may have fewer dimensions than the original.

Normalize Parameters

output_type: The element type for the restored raster. Rounds to the nearest integer if an Integer type is given.

PCA Parameters

output_type: The element type for the restored raster. Rounds to the nearest integer if an Integer type is given.

RemoteSensingToolbox.Transformations.transform — Method

transform(transformation::AbstractTransformation, raster; kwargs...)
transform(transformation::Normalize, raster)
transform(transformation::PCA, raster; output_type=Float32)

Performs the fitted transformation to the provided AbstractRaster or AbstractRasterStack.

Parameters

transformation: The transformation we want to apply.

raster: The AbstractRaster or AbstractRasterStack on which to perform the given transformation.

Supported Transformations

Normalize: Scales and shifts the data so that each band has a mean of 0 and a standard deviation of 1.

PCA: Rotates each pixel into a new orthogonal color-space, which may have fewer dimensions than the original.

Normalize Parameters

output_type: The element type for the transformed raster. Rounds to the nearest integer if an Integer type is given.

PCA Parameters

output_type: The element type for the transformed raster. Rounds to the nearest integer if an Integer type is given.