RemoteSensingToolbox

RemoteSensingToolbox is a pure Julia package intended to provide a collection of tools for visualizing, manipulating, and interpreting remotely sensed imagery.

Bandsets

Bandsets are julia types that encode the sensor-specific information needed for many methods in RemoteSensingToolbox to work without the need for tedious details provided by the end user. Bandsets are provided for several common sensors including Sentinel-2, Landsat 7, and Landsat 8. Unsupported sensors can be added by defining a custom type, which should be a sub-type of AbstractBandset, and implementing the following interface:

Method	Description	Required

bands	Return the band names in order from shortest to longest wavelength.	yes

wavelengths	Return the central wavelengths for all bands from shortest to longest.	yes

blue	Return the blue band for the given sensor.	yes

green	Return the green band for the given sensor.	yes

red	Return the red band for the given sensor.	yes

nir	Return the nir band for the given sensor.	yes

swir1	Return the swir1 band for the given sensor.	yes

swir2	Return the swir2 band for the given sensor.	yes

parse_band	Parses the band from a given filename. Used by read_bands.	no

read_qa	Reads the QA or scene classification file from the provided file or directory.	no

dn_to_reflectance	Decodes digital numbers to reflectance.	no

RemoteSensingToolbox.Bandsets.AbstractBandset — Type

The supertype of all sensor types. Provides sensor-specific information to many RemoteSensingToolbox methods.

Example Minimal Implementation

struct Landsat8 <: AbstractBandset end

bands(::Type{Landsat8}) = [:B1, :B2, :B3, :B4, :B5, :B6, :B7]

wavelengths(::Type{Landsat8}) = [443, 483, 560, 660, 865, 1650, 2220]

blue(::Type{Landsat8}) = :B2

green(::Type{Landsat8}) = :B3

red(::Type{Landsat8}) = :B4

nir(::Type{Landsat8}) = :B5

swir1(::Type{Landsat8}) = :B6

swir2(::Type{Landsat8}) = :B7

RemoteSensingToolbox.Bandsets.DESIS — Type

Implements the AbstractBandset interface for Landsat 7.

RemoteSensingToolbox.Bandsets.Landsat8 — Type

Implements the AbstractBandset interface for Landsat 8.

RemoteSensingToolbox.Bandsets.Landsat7 — Type

Implements the AbstractBandset interface for Landsat 7.

RemoteSensingToolbox.Bandsets.Sentinel2 — Type

Implements the AbstractBandset interface for Sentinel 2.

RemoteSensingToolbox.Bandsets.red — Function

red(::Type{AbstractBandset})
red(raster, ::Type{AbstractBandset})

Return the red band for the given sensor.

RemoteSensingToolbox.Bandsets.green — Function

green(::Type{AbstractBandset})
green(raster, ::Type{AbstractBandset})

Return the green band for the given sensor.

RemoteSensingToolbox.Bandsets.blue — Function

blue(::Type{AbstractBandset})
blue(raster, ::Type{AbstractBandset})

Return the blue band for the given sensor.

RemoteSensingToolbox.Bandsets.nir — Function

nir(::Type{AbstractBandset})
nir(raster, ::Type{AbstractBandset})

Return the nir band for the given sensor.

RemoteSensingToolbox.Bandsets.swir1 — Function

swir1(::Type{AbstractBandset})
swir1(raster, ::Type{AbstractBandset})

Return the swir1 band for the given sensor.

RemoteSensingToolbox.Bandsets.swir2 — Function

swir2(::Type{AbstractBandset})
swir2(raster, ::Type{AbstractBandset})

Return the swir2 band for the given sensor.

RemoteSensingToolbox.Bandsets.bands — Function

bands(::Type{AbstractBandset})

Return the band names in order from shortest to longest wavelength.

RemoteSensingToolbox.Bandsets.wavelengths — Function

wavelengths(::Type{AbstractBandset})

Return the central wavelengths for all bands in order from shortest to longest.

RemoteSensingToolbox.Bandsets.wavelength — Function

wavelength(::Type{AbstractBandset}, band::Symbol)

Return the central wavelength for the corresponding band.

RemoteSensingToolbox.Bandsets.parse_band — Function

parse_band(::Type{AbstractBandset}, filename::String)

Parses the band name from the given file path.

Returns either the band as a Symbol or nothing if no band could be parsed.

If the file is a multi-band raster, returns the names of all bands as a Vector{Symbol}.

RemoteSensingToolbox.Bandsets.read_bands — Function

function read_bands(bandset::Type{AbstractBandset}, dir::String)

Read the bands from the given directory into a RasterStack.

Parameters

bandset: The sensor type to which the bands belong.

dir: The directory in which the bands can be read.

Example

julia> landsat = read_bands(Landsat8, "data/LC08_L2SP_043024_20200802_20200914_02_T1/")
RasterStack with dimensions: 
  X Projected{Float64} LinRange{Float64}(493785.0, 728385.0, 7821) ForwardOrdered Regular Points crs: WellKnownText,
  Y Projected{Float64} LinRange{Float64}(5.84638e6, 5.60878e6, 7921) ReverseOrdered Regular Points crs: WellKnownText
and 7 layers:
  :B1 UInt16 dims: X, Y (7821×7921)
  :B2 UInt16 dims: X, Y (7821×7921)
  :B3 UInt16 dims: X, Y (7821×7921)
  :B4 UInt16 dims: X, Y (7821×7921)
  :B5 UInt16 dims: X, Y (7821×7921)
  :B6 UInt16 dims: X, Y (7821×7921)
  :B7 UInt16 dims: X, Y (7821×7921)

RemoteSensingToolbox.Bandsets.read_qa — Function

read_qa(bandset::Type{AbstractBandSet}, src::String)

Read and decode the quality assurance mask for the given AbstractBandset.

Parameters

bandset: A subtype of AbstractBandset.

src: Either a directory containing the quality assurance mask named according to standard conventions or the file itself.

Returns

The decoded quality assurance mask as a RasterStack. Masked values are encoded as 1, non-masked values as 0, and missing values as 255.

Example

julia> qa = read_qa(Landsat8, "data/LC08_L2SP_043024_20200802_20200914_02_T1/")
RasterStack with dimensions: 
  X Projected{Float64} LinRange{Float64}(493785.0, 728385.0, 7821) ForwardOrdered Regular Points crs: WellKnownText,
  Y Projected{Float64} LinRange{Float64}(5.84638e6, 5.60878e6, 7921) ReverseOrdered Regular Points crs: WellKnownText
and 7 layers:
  :dilated_cloud UInt8 dims: X, Y (7821×7921)
  :cirrus        UInt8 dims: X, Y (7821×7921)
  :cloud         UInt8 dims: X, Y (7821×7921)
  :cloud_shadow  UInt8 dims: X, Y (7821×7921)
  :snow          UInt8 dims: X, Y (7821×7921)
  :clear         UInt8 dims: X, Y (7821×7921)
  :water         UInt8 dims: X, Y (7821×7921)

RemoteSensingToolbox.Bandsets.dn_to_reflectance — Function

dn_to_reflectance(stack::AbstractRasterStack, bandset::Type{AbstractBandset}; clamp_values=false)

Transform the raster from Digital Numbers (DN) to reflectance.

Parameters

stack: The AbstractRasterStack to be converted to reflectance.

bandset: A subtype of AbstractBandset.

clamp_values: Indicates whether to clamp reflectances into the range (0.0, 1.0] (default = false).

Example

landsat = read(Landsat8, "LC08_L2SP_043024_20200802_20200914_02_T1/")
landsat_sr = dn_to_reflectance(landsat, Landsat8)

Visualization

RemoteSensingToolbox.Agriculture — Type

Agriculture band composite.

RemoteSensingToolbox.ColorInfrared — Type

Color infrared band composite.

RemoteSensingToolbox.Geology — Type

Geology band composite.

RemoteSensingToolbox.SWIR — Type

SWIR band composite.

RemoteSensingToolbox.TrueColor — Type

True color band composite.

RemoteSensingToolbox._linear_stretch — Method

Adjust image histogram by performing a linear stretch to squeeze all values between the percentiles lower and upper into the range [0,1].

RemoteSensingToolbox._raster_to_image — Method

Turn a raster into an image compatible with Images.jl.

RemoteSensingToolbox.visualize — Method

visualize(g::AbstractRaster; lower=0.02, upper=0.98)
visualize(r::AbstractRaster, g::AbstractRaster, b::AbstractRaster; lower=0.02, upper=0.98)
visualize(img::AbstractBandSet, ::Type{TrueColor{AbstractBandset}}; kwargs...)
visualize(img::AbstractBandSet, ::Type{ColorInfrared{AbstractBandset}}; kwargs...)
visualize(img::AbstractBandSet, ::Type{SWIR{AbstractBandset}}; kwargs...)
visualize(img::AbstractBandSet, ::Type{Agriculture{AbstractBandset}}; kwargs...)
visualize(img::AbstractBandSet, ::Type{Geology{AbstractBandset}}; kwargs...)

Visualize a satellite image after applying a histogram stretch. Returns either an RGB or grayscale image compatible with the Images.jl ecosystem.

A number of band combinations are supported for types implementing the AbstractBandSet interface.

Example 1

landsat = read(Landsat8, "LC08_L2SP_043024_20200802_20200914_02_T1/")
img = mndwi(landsat, Landsat8) |> visualize
save("mndwi.png", img)

Example 2

landsat = read(Landsat8, "LC08_L2SP_043024_20200802_20200914_02_T1/")
img = visualize(landsat, TrueColor{Landsat8}; upper=0.90)
save("truecolor.png", img)

Preprocessing

RemoteSensingToolbox.create_tiles — Method

create_tiles(raster, tile::Tuple{Int,Int}; stride=tile)

Slice the given raster into tiles with size tile.

Parameters

raster: The raster to be cut into tiles.

tile: The size of the generated tiles in terms of width x height.

stride: The distance between the top-left corner of each tile. Is equal to tile by default, which produces non-overlapping tiles.

RemoteSensingToolbox.mask_pixels! — Method

mask_pixels!(raster, mask; invert_mask=false)

Drop pixels from a raster according to a given mask. The mask and raster must have the same extent and size.

Parameters

raster: The raster to be masked.

mask: A mask defining which pixels we want to drop. By default, we drop pixels corresponding to mask values of 1.

invert_mask: Treat mask values of 1 as 0 and vice-versa.

RemoteSensingToolbox.mask_pixels — Method

mask_pixels(raster, mask; invert_mask=false)

Drop pixels from a raster according to a given mask. The mask and raster must have the same extent and size.

Parameters

raster: The raster to be masked.

mask: A mask defining which pixels we want to drop. By default, we drop pixels corresponding to mask values of 1.

invert_mask: Treat mask values of 1 as 0 and vice-versa.

RemoteSensingToolbox.tocube — Method

tocube(rs::RasterStack; layers=names(rs))
tocube(rs::AbstractBandset; layers=names(rs))

Transform the multi-layer RasterStack to a multi-band raster.

Parameters

X: The RasterStack or AbstractBandset to be transformed into a multi-band raster.

layers: The layers to include in the new raster.

Example

julia> landsat = Landsat8("LC08_L2SP_043024_20200802_20200914_02_T1");
julia> tocube(landsat)
7821×7921×7 Raster{Float32,3} B1 with dimensions: 
  X Projected{Float64} LinRange{Float64}(493785.0, 728385.0, 7821) ForwardOrdered Regular Points crs: WellKnownText,
  Y Projected{Float64} LinRange{Float64}(5.84638e6, 5.60878e6, 7921) ReverseOrdered Regular Points crs: WellKnownText,
  Band Categorical{Int64} 1:7 ForwardOrdered
extent: Extent(X = (493785.0, 728385.0), Y = (5.608785e6, 5.846385e6), Band = (1, 7))
missingval: 0.0f0
crs: PROJCS["WGS 84 / UTM zone 11N",GEOGCS["WGS 84",DATUM["WGS_1984",SPHEROID["WGS 84",6378137,298.257223563,AUTHORITY["EPSG","7030"]],AUTHORITY["EPSG","6326"]],PRIMEM["Greenwich",0,AUTHORITY["EPSG","8901"]],UNIT["degree",0.0174532925199433,AUTHORITY["EPSG","9122"]],AUTHORITY["EPSG","4326"]],PROJECTION["Transverse_Mercator"],PARAMETER["latitude_of_origin",0],PARAMETER["central_meridian",-117],PARAMETER["scale_factor",0.9996],PARAMETER["false_easting",500000],PARAMETER["false_northing",0],UNIT["metre",1,AUTHORITY["EPSG","9001"]],AXIS["Easting",EAST],AXIS["Northing",NORTH],AUTHORITY["EPSG","32611"]]
parent:
[:, :, 1]
           5.84638e6  5.84636e6  5.84632e6  5.8463e6  5.84626e6  …  5.60894e6  5.6089e6  5.60888e6  5.60884e6  5.60882e6  5.60878e6
 493785.0  0.0        0.0        0.0        0.0       0.0           0.0        0.0       0.0        0.0        0.0        0.0
      ⋮                                               ⋮          ⋱                                             ⋮          
 728355.0  0.0        0.0        0.0        0.0       0.0           0.0        0.0       0.0        0.0        0.0        0.0
 728385.0  0.0        0.0        0.0        0.0       0.0           0.0        0.0       0.0        0.0        0.0        0.0
[and 6 more slices...]

Land Cover Indices

RemoteSensingToolbox.mndwi — Method

mndwi(green::AbstractRaster, swir::AbstractRaster)
mndwi(stack::AbstractRasterStack, ::Type{AbstractBandset})

Compute the Modified Normalised Difference Water Index (Xu 2006).

MNDWI = (green - swir) / (green + swir)

RemoteSensingToolbox.nbri — Method

nbri(nir::AbstractRaster, swir2::AbstractRaster)
nbri(stack::AbstractRasterStack, ::Type{AbstractBandset})

Compute the Normalized Burn Ratio Index.

NBRI is used to emphasize burned areas.

NBRI = (nir - swir2) / (nir + swir2)

RemoteSensingToolbox.ndbi — Method

ndbi(swir1::AbstractRaster, nir::AbstractRaster)
ndbi(stack::AbstractRasterStack, ::Type{AbstractBandset})

Compute the The Normalized Difference Built-up Index

NDBI is used to emphasize urban and built-up areas.

NDBI = (swir1 - nir) / (swir1 + nir)

RemoteSensingToolbox.ndmi — Method

ndmi(nir::AbstractRaster, swir1::AbstractRaster)
ndmi(stack::AbstractRasterStack, ::Type{AbstractBandset})

Compute the Normalized Difference Moisture Index.

NDMI is sensitive to the moisture levels in vegetation. It is used to monitor droughts and fuel levels in fire-prone areas.

NDMI = (nir - swir1) / (nir + swir1)

RemoteSensingToolbox.ndvi — Method

ndvi(nir::AbstractRaster, red::AbstractRaster)
ndvi(stack::AbstractRasterStack, ::Type{AbstractBandset})

Compute the Normalized Difference Vegetation Index.

NDVI = (nir - red) / (nir + red)

RemoteSensingToolbox.ndwi — Method

ndwi(green::AbstractRaster, nir::AbstractRaster)
ndwi(stack::AbstractRasterStack, ::Type{AbstractBandset})

Compute the Normalized Difference Water Index (McFeeters 1996).

NDWI = (green - nir) / (green + nir)

RemoteSensingToolbox.savi — Method

savi(nir::AbstractRaster, red::AbstractRaster; L=0.33)
savi(stack::AbstractRasterStack, ::Type{AbstractBandset}; L=0.33)

Compute the Soil Adjusted Vegetation Index (Huete 1988).

SAVI is a vegetation index which attempts to minimize soil brightness influences by introducing a soil-brightness correction factor (L).

L represents the amount of green vegetation cover, which is set to 0.33 by default.

SAVI = ((nir - red) / (nir + red + L)) * (1 + L)

Spectral Analysis

RemoteSensingToolbox.Spectral.labelled_signatures — Method

labelled_signatures(rs::AbstractSensor, shp::DataFrame, label::Symbol)
labelled_signatures(rs::RasterStack, shp::DataFrame, label::Symbol)

Extract signatures from the given RasterStack or AbstractSensor within regions specified by a given shapefile.

Parameters

rs: The RasterStack or AbstractSensor from which to extract spectral signatures.

shp: A shapefile stored as a DataFrame with a :geometry column storing a GeoInterface.jl compatible geometry and a label column indicating the land cover type.

label: The column in shp in which the land cover class is stored.

Returns

A DataFrame consisting of rows for each extracted signature and columns storing the respective bands and land cover type.

Example

julia> landsat = Landsat8("data/LC08_L2SP_043024_20200802_20200914_02_T1/") |> dn_to_reflectance;

julia> shp = Shapefile.Table("data/landcover/landcover.shp") |> DataFrame;

julia> labelled_signatures(landsat, shp, :cover)
2860×8 DataFrame
  Row │ B1         B2         B3         B4         B5        B6        B7        label      
      │ Float32    Float32    Float32    Float32    Float32   Float32   Float32   String     
──────┼──────────────────────────────────────────────────────────────────────────────────────
    1 │ 0.0514875  0.073845   0.119742   0.128817   0.2466    0.256005  0.221905  Built Up
    2 │ 0.0759625  0.10228    0.152963   0.170865   0.27575   0.300115  0.26387   Built Up
    3 │ 0.113747   0.134263   0.212747   0.237497   0.32756   0.32041   0.28851   Built Up
    4 │ 0.09359    0.117955   0.18522    0.19743    0.302453  0.3148    0.277125  Built Up
    5 │ 0.0793725  0.104507   0.133658   0.1377     0.311555  0.2873    0.23824   Built Up
    6 │ 0.0641375  0.086      0.11724    0.115315   0.353272  0.291755  0.221878  Built Up
    7 │ 0.240688   0.271735   0.254245   0.299648   0.376895  0.500067  0.460797  Built Up
    8 │ 0.06554    0.0865225  0.14265    0.16245    0.272478  0.283697  0.238515  Built Up
    9 │ 0.0979625  0.131732   0.190747   0.230402   0.312793  0.369112  0.327148  Built Up
  ⋮   │     ⋮          ⋮          ⋮          ⋮         ⋮         ⋮         ⋮          ⋮
 2853 │ 0.0355925  0.0422475  0.0688675  0.0711225  0.24385   0.287245  0.180792  Bare Earth
 2854 │ 0.0494525  0.05674    0.0867975  0.09612    0.277703  0.311665  0.192728  Bare Earth
 2855 │ 0.0473075  0.053      0.0815725  0.0860825  0.264998  0.291672  0.183075  Bare Earth
 2856 │ 0.04673    0.05267    0.08039    0.083195   0.25716   0.285787  0.184175  Bare Earth
 2857 │ 0.0465925  0.05289    0.0809675  0.0854225  0.268627  0.299345  0.189565  Bare Earth
 2858 │ 0.0471975  0.052835   0.08094    0.0856975  0.266097  0.302783  0.189097  Bare Earth
 2859 │ 0.0453275  0.050745   0.075055   0.0795925  0.247095  0.293323  0.185962  Bare Earth
 2860 │ 0.036335   0.0434025  0.06884    0.070985   0.2422    0.282845  0.179995  Bare Earth
                                                                            2843 rows omitted

RemoteSensingToolbox.Spectral.plot_signatures! — Method

plot_signatures!(::Type{AbstractBandset}, ax::Axis, rs::RasterStack, shp::DataFrame, label::Symbol; colors=wong_colors())

Plot spectral signatures for each land cover type specified in a given shapefile by mutating a Makie.Axis object.

Parameters

bandset: The sensor type to which the signatures belong.

ax: The Makie.Axis into which we want to draw our plot.

rs: The RasterStack or AbstractSensor from which to extract spectral signatures.

shp: A shapefile stored as a DataFrame with a :geometry column storing a GeoInterface.jl compatible geometry and a label column indicating the land cover type.

label: The column in shp in which the land cover class is stored.

colors: The color scheme used by the plot.

Example

# Read Landsat And Convert DNs To Reflectance
landsat = @pipe read_bands(Landsat8, "data/LC08_L2SP_043024_20200802_20200914_02_T1/") |> dn_to_reflectance(_, Landsat8)

# Load Shapefile
shp = Shapefile.Table("data/landcover/landcover.shp") |> DataFrame

# Create Axes
fig = Figure();
ax1 = Axis(fig[1,1]);
ax2 = Axis(fig[2,1]);

# Plot Signatures
plot_signatures!(Landsat8, ax1, landsat, shp, :C_name; colors=cgrad(:tab10))
plot_signatures!(Landsat8, ax2, landsat, shp, :C_name; colors=cgrad(:tab10))

# Add Legend
Legend(fig[1,2], ax1);
Legend(fig[2,2], ax2);

RemoteSensingToolbox.Spectral.plot_signatures — Method

plot_signatures(bandset::Type{AbstractBandset}, rs::RasterStack, shp::DataFrame, label::Symbol; colors=wong_colors())

Plot spectral signatures for each land cover type specified in a given shapefile.

Parameters

bandset: The sensor type to which the signatures belong.

rs: The RasterStack or AbstractSensor from which to extract spectral signatures.

shp: A shapefile stored as a DataFrame with a :geometry column storing a GeoInterface.jl compatible geometry and a label column indicating the land cover type.

label: The column in shp in which the land cover class is stored.

colors: The color scheme used by the plot.

Example

# Read Landsat And Convert DNs To Reflectance
landsat = @pipe read_bands(Landsat8, "data/LC08_L2SP_043024_20200802_20200914_02_T1/") |> dn_to_reflectance(_, Landsat8)

# Load Shapefile
shp = Shapefile.Table("data/landcover/landcover.shp") |> DataFrame

# Plot Signatures
plot_signatures(Landsat8, landsat, shp, :C_name; colors=cgrad(:tab10))

Transformations

RemoteSensingToolbox.Transformations.AbstractTransformation — Type

The supertype of all transformations. Subtypes are expected to implement the fit and transform methods.

RemoteSensingToolbox.Transformations.Normalize — Type

A struct for storing the parameters necessary to perform a normalization transformation.

RemoteSensingToolbox.Transformations.PCA — Type

A struct for storing the parameters necessary to perform a PCA transformation.

RemoteSensingToolbox.Transformations.fit_transform — Method

fit_transform(transformation::Type{AbstractTransformation}, raster; kwargs...)
fit_transform(transformation::Type{Normalize}, raster)
fit_transform(transformation::Type{PCA}, raster; components=nbands(raster), method=:cov, stats_fraction=1.0)

Fit the specified transformation to the given AbstractRasterStack or AbstractRaster.

Parameters

transformation: The transformation we want to fit.

raster: The AbstractRaster or AbstractRasterStack on which to fit the specified transformation.

Supported Transformations

Normalize: Scales and shifts the data so that each band has a mean of 0 and a standard deviation of 1.

PCA: Rotates each pixel into a new orthogonal color-space, which may have fewer dimensions than the original.

PCA Parameters

components: The number of principal components to use.

method: One of either :cov or :cor, depending on whether we want to run PCA on the covariance or the correlation matrix.

stats_fraction: The fraction of pixels to use in the calculation. Setting stats_fraction < 1 will produce faster but less accurate results.

RemoteSensingToolbox.Transformations.inverse_transform — Method

inverse_transform(transformation::AbstractTransformation, raster)
inverse_transform(transformation::Normalize, raster; output_type=Float32)
inverse_transform(transformation::PCA, raster; output_type=Float32)

Undo a previously applied transformation.

Parameters

transformation: Some previously applied transformation that we want to reverse.

raster: A previously transformed AbstractRaster or AbstractRasterStack from which we want to recover the original.

Supported Transformations

Normalize: Scales and shifts the data so that each band has a mean of 0 and a standard deviation of 1.

PCA: Rotates each pixel into a new orthogonal color-space, which may have fewer dimensions than the original.

Normalize Parameters

output_type: The element type for the restored raster. Rounds to the nearest integer if an Integer type is given.

PCA Parameters

output_type: The element type for the restored raster. Rounds to the nearest integer if an Integer type is given.

RemoteSensingToolbox.Transformations.transform — Method

transform(transformation::AbstractTransformation, raster; kwargs...)
transform(transformation::Normalize, raster)
transform(transformation::PCA, raster; output_type=Float32)

Performs the fitted transformation to the provided AbstractRaster or AbstractRasterStack.

Parameters

transformation: The transformation we want to apply.

raster: The AbstractRaster or AbstractRasterStack on which to perform the given transformation.

Supported Transformations

Normalize: Scales and shifts the data so that each band has a mean of 0 and a standard deviation of 1.

PCA: Rotates each pixel into a new orthogonal color-space, which may have fewer dimensions than the original.

Normalize Parameters

output_type: The element type for the transformed raster. Rounds to the nearest integer if an Integer type is given.

PCA Parameters

output_type: The element type for the transformed raster. Rounds to the nearest integer if an Integer type is given.